+++ python/dsbox/template/dsbox_templates/dsbox_image.py
@@ -1,0 +1,226 @@
+from d3m.metadata.problem import TaskType, TaskSubtype
+from .template_steps import TemplateSteps
+from dsbox.template.template import DSBoxTemplate
+import numpy as np
+
+
+
+################################################################################################################
+#####################################   ImageProblemsTemplates   ###############################################
+################################################################################################################
+
+
+class TA1VggImageProcessingRegressionTemplate(DSBoxTemplate):
+    def __init__(self):
+        DSBoxTemplate.__init__(self)
+        self.template = {
+            "name": "TA1VggImageProcessingRegressionTemplate",
+            "taskType": TaskType.REGRESSION.name,
+            # See TaskType, range include 'CLASSIFICATION', 'CLUSTERING', 'COLLABORATIVE_FILTERING', 'COMMUNITY_DETECTION', 'GRAPH_CLUSTERING', 'GRAPH_MATCHING', 'LINK_PREDICTION', 'REGRESSION', 'TIME_SERIES_FORECASTING', 'VERTEX_NOMINATION'
+            "taskSubtype": {TaskSubtype.UNIVARIATE.name, TaskSubtype.MULTIVARIATE.name},
+            "inputType": "image",  # See SEMANTIC_TYPES.keys() for range of values
+            "output": "regressor_step",  # Name of the final step generating the prediction
+            "target": "extract_target_step",  # Name of the step generating the ground truth
+            "steps": [
+                {
+                    "name": "denormalize_step",
+                    "primitives": ["d3m.primitives.dsbox.Denormalize"],
+                    "inputs": ["template_input"]
+                },
+                {
+                    "name": "to_dataframe_step",
+                    "primitives": ["d3m.primitives.datasets.DatasetToDataFrame"],
+                    "inputs": ["denormalize_step"]
+                },
+                # read Y value
+                {
+                    "name": "extract_target_step",
+                    "primitives": [{
+                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
+                        "hyperparameters":
+                            {
+                                'semantic_types': ('https://metadata.datadrivendiscovery.org/types/TrueTarget',),
+                                'use_columns': (),
+                                'exclude_columns': ()
+                            }
+                        }],
+                    "inputs": ["to_dataframe_step"]
+                },
+                # {
+                #     "name": "column_parser_step",
+                #     "primitives": ["d3m.primitives.data.ColumnParser"],
+                #     "inputs": ["to_dataframe_step"]
+                # },
+                # read X value
+                {
+                    "name": "dataframe_to_tensor",
+                    "primitives": ["d3m.primitives.dsbox.DataFrameToTensor"],
+                    "inputs": ["to_dataframe_step"]
+                },
+                {
+                    "name": "feature_extraction",
+                    "primitives": ["d3m.primitives.dsbox.ResNet50ImageFeature"],
+                    # or "primitives": ["d3m.primitives.dsbox.Vgg16ImageFeature"],
+                    "inputs": ["dataframe_to_tensor"]
+                },
+                {
+                    "name": "PCA_step",
+                    "primitives": ["d3m.primitives.sklearn_wrap.SKPCA"],
+                    "inputs": ["feature_extraction"]
+                },
+
+                {
+                    "name": "regressor_step",
+                    "primitives": ["d3m.primitives.sklearn_wrap.SKRandomForestRegressor"],
+                    "inputs": ["PCA_step", "extract_target_step"]
+                },
+            ]
+        }
+
+    def importance(datset, problem_description):
+        return 7
+
+
+class TA1DefaultImageProcessingRegressionTemplate(DSBoxTemplate):
+    def __init__(self):
+        DSBoxTemplate.__init__(self)
+        self.template = {
+            "name": "Default_image_processing_regression_template",
+            "taskType": TaskType.REGRESSION.name,
+            # See TaskType, range include 'CLASSIFICATION', 'CLUSTERING', 'COLLABORATIVE_FILTERING',
+            # 'COMMUNITY_DETECTION', 'GRAPH_CLUSTERING', 'GRAPH_MATCHING', 'LINK_PREDICTION',
+            # 'REGRESSION', 'TIME_SERIES_FORECASTING', 'VERTEX_NOMINATION'
+            "taskSubtype": {TaskSubtype.UNIVARIATE.name, TaskSubtype.MULTIVARIATE.name},
+            "inputType": "image",  # See SEMANTIC_TYPES.keys() for range of values
+            "output": "regressor_step",  # Name of the final step generating the prediction
+            "target": "extract_target_step",  # Name of the step generating the ground truth
+            "steps": [
+
+                {
+                    "name": "denormalize_step",
+                    "primitives": ["d3m.primitives.dsbox.Denormalize"],
+                    "inputs": ["template_input"]
+                },
+                {
+                    "name": "to_dataframe_step",
+                    "primitives": ["d3m.primitives.datasets.DatasetToDataFrame"],
+                    "inputs": ["denormalize_step"]
+                },
+                # read Y value
+                {
+                    "name": "extract_target_step",
+                    "primitives": [{
+                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
+                        "hyperparameters":
+                            {
+                                'semantic_types': ('https://metadata.datadrivendiscovery.org/types/TrueTarget',),
+                                'use_columns': (),
+                                'exclude_columns': ()
+                            }
+                        }],
+                    "inputs": ["to_dataframe_step"]
+                },
+                # {
+                #     "name": "column_parser_step",
+                #     "primitives": ["d3m.primitives.data.ColumnParser"],
+                #     "inputs": ["to_dataframe_step"]
+                # },
+                # read X value
+                {
+                    "name": "dataframe_to_tensor",
+                    "primitives": ["d3m.primitives.dsbox.DataFrameToTensor"],
+                    "inputs": ["to_dataframe_step"]
+                },
+                {
+                    "name": "feature_extraction",
+                    "primitives": ["d3m.primitives.dsbox.ResNet50ImageFeature"],
+                    #"primitives": ["d3m.primitives.dsbox.Vgg16ImageFeature"],
+                    "inputs": ["dataframe_to_tensor"]
+                },
+                {
+                    "name": "PCA_step",
+                    "primitives": ["d3m.primitives.sklearn_wrap.SKPCA"],
+                    "inputs": ["feature_extraction"]
+                },
+
+                {
+                    "name": "regressor_step",
+                    "primitives": ["d3m.primitives.sklearn_wrap.SKRandomForestRegressor"],
+                    "inputs": ["PCA_step", "extract_target_step"]
+                },
+            ]
+        }
+
+    def importance(datset, problem_description):
+        return 7
+
+
+class DefaultImageProcessingRegressionTemplate(DSBoxTemplate):
+    def __init__(self):
+        DSBoxTemplate.__init__(self)
+        self.template = {
+            "name": "Default_image_processing_regression_template",
+            "taskType": TaskType.REGRESSION.name,
+            # See TaskType, range include 'CLASSIFICATION', 'CLUSTERING', 'COLLABORATIVE_FILTERING', 'COMMUNITY_DETECTION', 'GRAPH_CLUSTERING', 'GRAPH_MATCHING', 'LINK_PREDICTION', 'REGRESSION', 'TIME_SERIES_FORECASTING', 'VERTEX_NOMINATION'
+            "taskSubtype": {TaskSubtype.UNIVARIATE.name, TaskSubtype.MULTIVARIATE.name},
+            "inputType": "image",  # See SEMANTIC_TYPES.keys() for range of values
+            "output": "regressor_step",  # Name of the final step generating the prediction
+            "target": "extract_target_step",  # Name of the step generating the ground truth
+            "steps": [
+
+                {
+                    "name": "denormalize_step",
+                    "primitives": ["d3m.primitives.dsbox.Denormalize"],
+                    "inputs": ["template_input"]
+                },
+                {
+                    "name": "to_dataframe_step",
+                    "primitives": ["d3m.primitives.datasets.DatasetToDataFrame"],
+                    "inputs": ["denormalize_step"]
+                },
+                # read Y value
+                {
+                    "name": "extract_target_step",
+                    "primitives": [{
+                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
+                        "hyperparameters":
+                            {
+                                'semantic_types': ('https://metadata.datadrivendiscovery.org/types/TrueTarget',),
+                                'use_columns': (),
+                                'exclude_columns': ()
+                            }
+                        }],
+                    "inputs": ["to_dataframe_step"]
+                },
+                # {
+                #     "name": "column_parser_step",
+                #     "primitives": ["d3m.primitives.data.ColumnParser"],
+                #     "inputs": ["to_dataframe_step"]
+                # },
+                # read X value
+                {
+                    "name": "dataframe_to_tensor",
+                    "primitives": ["d3m.primitives.dsbox.DataFrameToTensor"],
+                    "inputs": ["to_dataframe_step"]
+                },
+                {
+                    "name": "feature_extraction",
+                    "primitives": ["d3m.primitives.dsbox.ResNet50ImageFeature"],
+                    "inputs": ["dataframe_to_tensor"]
+                },
+                {
+                    "name": "PCA_step",
+                    "primitives": ["d3m.primitives.sklearn_wrap.SKPCA"],
+                    "inputs": ["feature_extraction"]
+                },
+
+                {
+                    "name": "regressor_step",
+                    "primitives": ["d3m.primitives.sklearn_wrap.SKRandomForestRegressor"],
+                    "inputs": ["PCA_step", "extract_target_step"]
+                },
+            ]
+        }
+
+    def importance(datset, problem_description):
+        return 7
