+++ python/dsbox/template/dsbox_templates/dsbox_other.py
@@ -1,0 +1,406 @@
+from d3m.metadata.problem import TaskType, TaskSubtype
+from .template_steps import TemplateSteps
+from dsbox.template.template import DSBoxTemplate
+import numpy as np
+
+################################################################################################################
+#####################################   ObjectDetectionTemplates   #############################################
+################################################################################################################
+
+
+class TemporaryObjectDetectionTemplate(DSBoxTemplate):
+    def __init__(self):
+        DSBoxTemplate.__init__(self)
+        self.template = {
+            "name": "TemporaryObjectDetectionTemplate",
+            "taskType": TaskType.OBJECT_DETECTION.name,
+            # See TaskType, range include 'CLASSIFICATION', 'CLUSTERING', 'COLLABORATIVE_FILTERING',
+            # 'COMMUNITY_DETECTION', 'GRAPH_CLUSTERING', 'GRAPH_MATCHING', 'LINK_PREDICTION',
+            # 'REGRESSION', 'TIME_SERIES_FORECASTING', 'VERTEX_NOMINATION'
+            "taskSubtype": "NONE",
+            "inputType": {"table", "image"},  # See SEMANTIC_TYPES.keys() for range of values
+            "output": "model_step",  # Name of the final step generating the prediction
+            "target": "extract_target_step",  # Name of the step generating the ground truth
+            "steps": [
+                {
+                    "name": "denormalize_step",
+                    "primitives": ["d3m.primitives.dsbox.Denormalize"],
+                    "inputs": ["template_input"]
+                },
+                {
+                    "name": "to_dataframe_step",
+                    "primitives": ["d3m.primitives.datasets.DatasetToDataFrame"],
+                    "inputs": ["denormalize_step"]
+                },
+                # read X value
+                {
+                    "name": "extract_file_step",
+                    "primitives": [{
+                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
+                        "hyperparameters":
+                            {
+                                'semantic_types': (
+                                    'https://metadata.datadrivendiscovery.org/types/FileName',),
+                                'use_columns': (),
+                                'exclude_columns': ()
+                            }
+                    }],
+                    "inputs": ["to_dataframe_step"]
+                },
+                {
+                    "name": "to_tensor_step",
+                    "primitives": ["d3m.primitives.dsbox.DataFrameToTensor"],
+                    "inputs": ["extract_file_step"]
+                },
+                {
+                    "name": "image_processing_step",
+                    "primitives": ["d3m.primitives.dsbox.ResNet50ImageFeature"],
+                    "inputs": ["to_tensor_step"]
+                },
+                # read Y value
+                {
+                    "name": "extract_target_step",
+                    "primitives": [{
+                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
+                        "hyperparameters":
+                            {
+                                'semantic_types': (
+                                    'https://metadata.datadrivendiscovery.org/types/TrueTarget',),
+                                'use_columns': (),
+                                'exclude_columns': ()
+                            }
+                    }],
+                    "inputs": ["to_dataframe_step"]
+                },
+
+                {
+                    "name":"data_clean_step",
+                    "primitives": ["d3m.primitives.dsbox.CleaningFeaturizer"],
+                    "inputs": ["extract_target_step"]
+                },
+
+                {
+                    "name": "model_step",
+                    "primitives": ["d3m.primitives.sklearn_wrap.SKRandomForestRegressor"],
+                    "inputs": ["image_processing_step", "data_clean_step"]
+                },
+            ]
+        }
+
+    # @override
+    def importance(datset, problem_description):
+        return 7
+
+
+
+
+
+
+
+
+
+################################################################################################################
+#####################################   AudioClassificationTemplate   ##########################################
+################################################################################################################
+
+
+class BBNAudioClassificationTemplate(DSBoxTemplate):
+    def __init__(self):
+        DSBoxTemplate.__init__(self)
+        self.template = {
+            "name": "BBN_Audio_Classification_Template",
+            "taskType": {TaskType.CLASSIFICATION.name},
+            "taskSubtype": {TaskSubtype.MULTICLASS.name},
+            "inputType": "audio",
+            "output": "model_step",
+            "steps": [
+                {
+                    "name": "denormalize_step",
+                    "primitives": ["d3m.primitives.dsbox.Denormalize"],
+                    "inputs": ["template_input"]
+                },
+                {
+                    "name": "to_dataframe_step",
+                    "primitives": ["d3m.primitives.datasets.DatasetToDataFrame"],
+                    "inputs": ["denormalize_step"]
+                },
+                {
+                    "name": "readtarget_step",
+                    "primitives": [{
+                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
+                        "hyperparameters":
+                            {
+                                'semantic_types': ('https://metadata.datadrivendiscovery.org/types/TrueTarget',),
+                                'use_columns': (),
+                                'exclude_columns': ()
+                            }
+                        }],
+                    "inputs": ["to_dataframe_step"]
+                },
+                {
+                    "name": "readaudio_step",
+                    "primitives": [{
+                        "primitive": "d3m.primitives.bbn.time_series.AudioReader",
+                        "hyperparameters":
+                            {
+                                "read_as_mono": [(True)],
+                                "resampling_rate": [(16000.0)],
+                            }
+                    }],
+                    "inputs": ["template_input"]
+                },
+                {
+                    "name": "channel_step",
+                    "primitives": ["d3m.primitives.bbn.time_series.ChannelAverager"],
+                    "inputs": ["readaudio_step"]
+                },
+                {
+                    "name": "signaldither_step",
+                    "primitives": [{"primitive": "d3m.primitives.bbn.time_series.SignalDither",
+                                    "hyperparameters": {
+                                        "level": [(0.0001)],
+                                        "reseed": [(True)]
+                                    }
+                                    }],
+                    "inputs": ["channel_step"]
+                },
+                {
+                    "name": "signalframer_step",
+                    "primitives": [{"primitive": "d3m.primitives.bbn.time_series.SignalFramer",
+                                    "hyperparameters": {
+                                        "flatten_output": [(False)],
+                                        "frame_length_s": [(0.025)],
+                                        "frame_shift_s": [(0.01)]
+                                    }
+                                    }],
+                    "inputs": ["signaldither_step"]
+                },
+                {
+                    "name": "MFCC_step",
+                    "primitives": [{
+                        "primitive": "d3m.primitives.bbn.time_series.SignalMFCC",
+                        "hyperparameters": {
+                            "cep_lifter": [(22.0)],
+                            "frame_mean_norm": [(False)],
+                            "nfft": [(None)],
+                            "num_ceps": [(20)],
+                            "num_chans": [(20)],
+                            "preemcoef": [(None)],
+                            "use_power": [(False)]
+                        }
+                    }],
+                    "inputs": ["signalframer_step"]
+                },
+                {
+                    "name": "vectorextractor_step",
+                    "primitives": [{
+                        "primitive": "d3m.primitives.bbn.time_series.IVectorExtractor",
+                        "hyperparameters": {
+                            "gmm_covariance_type": [("diag")],
+                            "ivec_dim": [(100)],
+                            "max_gmm_iter": [(20)],
+                            "num_gauss": [(32)],
+                            "num_ivec_iter": [(7)]
+                        }
+                    }],
+                    "inputs": ["MFCC_step"]
+                },
+                {
+                    "name": "model_step",
+                    "primitives": [{
+                        "primitive": "d3m.primitives.bbn.sklearn_wrap.BBNMLPClassifier",
+                        "hyperparameters": {
+                            "activation": [("relu")],
+                            "add_index_columns": [(True)],
+                            "alpha": [(0.0001)],
+                            "beta_1": [(0.9)],
+                            "beta_2": [(0.999)],
+                            "early_stopping": [(True)],
+                            "epsilon": [(1e-8)],
+                            "exclude_columns": [([])],
+                            # "hidden_layer_sizes":[([30,30])],
+                            "learning_rate": [("constant")],
+                            "learning_rate_init": [(0.01)],
+                            "max_iter": [(200)],
+                            "return_result": [("replace")],
+                            "shuffle": [(True)],
+                            "solver": [("adam")],
+                            "tol": [(0.0001)],
+                            "use_columns": [([])],
+                            "use_semantic_types": [(False)],
+                            "warm_start": [(False)]
+                        }
+                    }],
+                    "inputs": ["vectorextractor_step", "readtarget_step"]
+
+                }
+            ]
+        }
+
+    def importance(datset, problem_description):
+        return 7
+
+
+################################################################################################################
+#####################################   SRIMeanbaselineTemplate   ##############################################
+################################################################################################################
+
+
+class SRIMeanBaselineTemplate(DSBoxTemplate):
+    def __init__(self):
+        DSBoxTemplate.__init__(self)
+        self.template = {
+            "name": "SRI_Mean_Baseline_Template",
+            "taskType": "NONE",
+            "taskSubtype": "NONE",
+            "inputType": "NONE",
+            "output": "model_step",
+            "steps": [
+                {
+                    "name": "model_step",
+                    "primitives": ["d3m.primitives.sri.baseline.MeanBaseline"],
+                    "inputs": ["template_input"]
+
+                }
+            ]
+        }
+
+    def importance(dataset, problem_description):
+        return 10
+
+
+################################################################################################################
+#####################################   ClusteringTemplate   ###################################################
+################################################################################################################
+
+
+class CMUClusteringTemplate(DSBoxTemplate):
+    def __init__(self):
+        DSBoxTemplate.__init__(self)
+        self.template = {
+            "name": "CMU_Clustering_Template",
+            "taskType": TaskType.CLUSTERING.name,
+            "taskSubtype": "NONE",
+            "inputType": "table",
+            "output": "model_step",
+            "steps": [
+                {
+                    "name": "denormalize_step",
+                    "primitives": ["d3m.primitives.dsbox.Denormalize"],
+                    "inputs": ["template_input"]
+                },
+                {
+                    "name": "to_dataframe_step",
+                    "primitives": ["d3m.primitives.datasets.DatasetToDataFrame"],
+                    "inputs": ["denormalize_step"]
+                },
+                {
+                    "name": "column_parser_step",
+                    "primitives": ["d3m.primitives.data.ColumnParser"],
+                    "inputs": ["to_dataframe_step"]
+                },
+                {
+                    "name": "extract_attribute_step",
+                    "primitives": [{
+                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
+                        "hyperparameters":
+                            {
+                                'semantic_types': (
+                                    'https://metadata.datadrivendiscovery.org/types/Attribute',),
+                                'use_columns': (),
+                                'exclude_columns': ()
+                            }
+                    }],
+                    "inputs": ["column_parser_step"]
+                },
+                {
+                    "name": "model_step",
+                    "primitives": [
+                        {
+                            "primitive": "d3m.primitives.cmu.fastlvm.GMM",
+                            "hyperparameters": {
+                                "k": [(4), (6), (8), (10), (12)]
+                            }
+                        }
+                    ],
+                    "inputs": ["extract_attribute_step"]
+                }
+            ]
+        }
+
+    def importance(datset, problem_description):
+        return 7
+
+
+################################################################################################################
+#####################################   VideoClassificationTemplate   ##########################################
+################################################################################################################
+
+
+class MichiganVideoClassificationTemplate(DSBoxTemplate):
+    def __init__(self):
+        DSBoxTemplate.__init__(self)
+        self.template = {
+            "name": "Michigan_Video_Classification_Template",
+            "taskType": TaskType.CLASSIFICATION.name,
+            # See TaskType, range include 'CLASSIFICATION', 'CLUSTERING', 'COLLABORATIVE_FILTERING',
+            # 'COMMUNITY_DETECTION', 'GRAPH_CLUSTERING', 'GRAPH_MATCHING', 'LINK_PREDICTION',
+            # 'REGRESSION', 'TIME_SERIES_FORECASTING', 'VERTEX_NOMINATION'
+            "taskSubtype": TaskSubtype.MULTICLASS.name,
+            "inputType": "video",  # See SEMANTIC_TYPES.keys() for range of values
+            "output": "model_step",  # Name of the final step generating the prediction
+            "target": "extract_target_step",  # Name of the step generating the ground truth
+            "steps": [
+                {
+                    "name": "denormalize_step",
+                    "primitives": ["d3m.primitives.dsbox.Denormalize"],
+                    "inputs": ["template_input"]
+                },
+                {
+                    "name": "to_dataframe_step",
+                    "primitives": ["d3m.primitives.datasets.DatasetToDataFrame"],
+                    "inputs": ["denormalize_step"]
+                },
+                {
+                    "name": "extract_target_step",
+                    "primitives": [{
+                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
+                        "hyperparameters":
+                            {
+                                'semantic_types': (
+                                    'https://metadata.datadrivendiscovery.org/types/TrueTarget',),
+                                'use_columns': (),
+                                'exclude_columns': ()
+                            }
+                    }],
+                    "inputs": ["to_dataframe_step"]
+                },
+                {
+                    "name": "read_video_step",
+                    "primitives": ["d3m.primitives.data.VideoReader"],
+                    "inputs": ["to_dataframe_step"]
+                },
+                {
+                    "name": "featurize_step",
+                    "primitives": ["d3m.primitives.spider.featurization.I3D"],
+                    "inputs": ["read_video_step"]
+
+                },
+                {
+                    "name": "convert_step",
+                    "primitives": ["d3m.primitives.data.NDArrayToDataFrame"],
+                    "inputs": ["featurize_step"]
+
+                },
+                {
+                    "name": "model_step",
+                    "primitives": ["d3m.primitives.classifier.RandomForest"],
+                    "inputs": ["convert_step", "extract_target_step"]
+                },
+            ]
+        }
+
+    # @override
+    def importance(datset, problem_description):
+        return 7
+
