+++ python/dsbox/template/library.py
@@ -8,9 +8,29 @@ from d3m.container.dataset import SEMANTIC_TYPES
 from d3m.metadata.problem import TaskType, TaskSubtype
 from d3m.container.list import List
 from dsbox.template.template import TemplatePipeline, DSBoxTemplate
-from .template_steps import TemplateSteps
+from dsbox.template.dsbox_templates.template_steps import TemplateSteps
 import copy
 
+
+from dsbox.template.dsbox_templates.dsbox_classification import *
+
+from dsbox.template.dsbox_templates.dsbox_regression import *
+
+from dsbox.template.dsbox_templates.dsbox_large_datasets import *
+
+from dsbox.template.dsbox_templates.dsbox_timeseries_forcasting import *
+
+from dsbox.template.dsbox_templates.dsbox_timeseries_general import *
+
+from dsbox.template.dsbox_templates.dsbox_image import *
+
+from dsbox.template.dsbox_templates.dsbox_text import *
+
+from dsbox.template.dsbox_templates.dsbox_graph import *
+
+from dsbox.template.dsbox_templates.dsbox_TA1 import *
+
+from dsbox.template.dsbox_templates.dsbox_other import *
 
 
 class TemplateDescription:
@@ -284,3176 +304,4 @@ class SemanticTypeDict(object):
 
 ################################################################################################################
 
-
-################################################################################################################
-##################################### General Classification Templates #########################################
-################################################################################################################
-
-################################################################################################
-# A classification template encompassing several algorithms
-################################################################################################
-class DefaultClassificationTemplate(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "default_classification_template",
-            "taskSubtype": {TaskSubtype.BINARY.name, TaskSubtype.MULTICLASS.name},
-            "taskType": TaskType.CLASSIFICATION.name,
-            # See TaskType, range include 'CLASSIFICATION', 'CLUSTERING', 'COLLABORATIVE_FILTERING',
-            # 'COMMUNITY_DETECTION', 'GRAPH_CLUSTERING', 'GRAPH_MATCHING', 'LINK_PREDICTION',
-            # 'REGRESSION', 'TIME_SERIES_FORECASTING', 'VERTEX_NOMINATION'
-            "inputType": "table",  # See SEMANTIC_TYPES.keys() for range of values
-            "output": "model_step",  # Name of the final step generating the prediction
-            "target": "extract_target_step",  # Name of the step generating the ground truth
-            "steps": TemplateSteps.dsbox_generic_steps() + [
-                {
-                    "name": "model_step",
-                    "runtime": {
-                        "cross_validation": 10,
-                        "stratified": True
-                    },
-                    "primitives": [
-                        {
-                            "primitive":
-                                "d3m.primitives.sklearn_wrap.SKRandomForestClassifier",
-                            "hyperparameters":
-                                {
-                                    'bootstrap': [True, False],
-                                    'max_depth': [15, 30, None],
-                                    'min_samples_leaf': [1, 2, 4],
-                                    'min_samples_split': [2, 5, 10],
-                                    'max_features': ['auto', 'sqrt'],
-                                    'n_estimators': [10, 50, 100],
-                                }
-                        },
-                        {
-                            "primitive":
-                                "d3m.primitives.sklearn_wrap.SKExtraTreesClassifier",
-                            "hyperparameters":
-                                {
-                                    'bootstrap': [True, False],
-                                    'max_depth': [15, 30, None],
-                                    'min_samples_leaf': [1, 2, 4],
-                                    'min_samples_split': [2, 5, 10],
-                                    'max_features': ['auto', 'sqrt'],
-                                    'n_estimators': [10, 50, 100],
-                                }
-                        },
-                        {
-                            "primitive":
-                                "d3m.primitives.sklearn_wrap.SKGradientBoostingClassifier",
-                            "hyperparameters":
-                                {
-                                    'max_depth': [2, 3, 4, 5],
-                                    'n_estimators': [50, 60, 80, 100],
-                                    'learning_rate': [0.1, 0.2, 0.4, 0.5],
-                                    'min_samples_split': [2, 3],
-                                    'min_samples_leaf': [1, 2],
-                                }
-                        },
-                    ],
-                    "inputs": ["data", "target"]
-                }
-            ]
-        }
-
-    # @override
-    def importance(datset, problem_description):
-        return 7
-
-
-# a template encompassing several NB methods
-class NaiveBayesClassificationTemplate(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "naive_bayes_classification_template",
-            "taskSubtype": {TaskSubtype.BINARY.name, TaskSubtype.MULTICLASS.name},
-            "taskType": TaskType.CLASSIFICATION.name,
-            # See TaskType, range include 'CLASSIFICATION', 'CLUSTERING', 'COLLABORATIVE_FILTERING',
-            # 'COMMUNITY_DETECTION', 'GRAPH_CLUSTERING', 'GRAPH_MATCHING', 'LINK_PREDICTION',
-            # 'REGRESSION', 'TIME_SERIES_FORECASTING', 'VERTEX_NOMINATION'
-            "inputType": "table",  # See SEMANTIC_TYPES.keys() for range of values
-            "output": "model_step",  # Name of the final step generating the prediction
-            "target": "extract_target_step",  # Name of the step generating the ground truth
-            "steps": TemplateSteps.dsbox_generic_steps() + [
-                {
-                    "name": "model_step",
-                    "runtime": {
-                        "cross_validation": 10,
-                        "stratified": True
-                    },
-                    "primitives": [
-                        {
-                            "primitive":
-                                "d3m.primitives.sklearn_wrap.SKBernoulliNB",
-                            "hyperparameters":
-                                {
-                                    'alpha': [0, .5, 1],
-                                }
-                        },
-                        {
-                            "primitive":
-                                "d3m.primitives.sklearn_wrap.SKGaussianNB",
-                            "hyperparameters":
-                                {
-                                }
-                        },
-                        {
-                            "primitive":
-                                "d3m.primitives.sklearn_wrap.SKMultinomialNB",
-                            "hyperparameters":
-                                {
-                                    'alpha': [0, .5, 1]
-                                }
-                        },
-                    ],
-                    "inputs": ["data", "target"]
-                }
-            ]
-        }
-
-    # @override
-    def importance(datset, problem_description):
-        return 7
-
-
-class RandomForestClassificationTemplate(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "random_forest_classification_template",
-            "taskSubtype": {TaskSubtype.BINARY.name, TaskSubtype.MULTICLASS.name},
-            "taskType": TaskType.CLASSIFICATION.name,
-            # See TaskType, range include 'CLASSIFICATION', 'CLUSTERING', 'COLLABORATIVE_FILTERING',
-            # 'COMMUNITY_DETECTION', 'GRAPH_CLUSTERING', 'GRAPH_MATCHING', 'LINK_PREDICTION',
-            # 'REGRESSION', 'TIME_SERIES_FORECASTING', 'VERTEX_NOMINATION'
-            "inputType": "table",  # See SEMANTIC_TYPES.keys() for range of values
-            "output": "model_step",  # Name of the final step generating the prediction
-            "target": "extract_target_step",  # Name of the step generating the ground truth
-            "steps": TemplateSteps.dsbox_generic_steps() + [
-                {
-                    "name": "model_step",
-                    "runtime": {
-                        "cross_validation": 10,
-                        "stratified": True
-                    },
-                    "primitives": [
-                        {
-                            "primitive":
-                                "d3m.primitives.sklearn_wrap.SKRandomForestClassifier",
-                            "hyperparameters":
-                                {
-                                    'bootstrap': [True, False],
-                                    'max_depth': [15, 30, None],
-                                    'min_samples_leaf': [1, 2, 4],
-                                    'min_samples_split': [2, 5, 10],
-                                    'max_features': ['auto', 'sqrt'],
-                                    'n_estimators': [10, 50, 100]
-                                }
-                        },
-                    ],
-                    "inputs": ["data", "target"]
-                }
-            ]
-        }
-
-    # @override
-    def importance(datset, problem_description):
-        return 7
-
-
-class ExtraTreesClassificationTemplate(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "extra_trees_classification_template",
-            "taskSubtype": {TaskSubtype.BINARY.name, TaskSubtype.MULTICLASS.name},
-            "taskType": TaskType.CLASSIFICATION.name,
-            # See TaskType, range include 'CLASSIFICATION', 'CLUSTERING', 'COLLABORATIVE_FILTERING',
-            # 'COMMUNITY_DETECTION', 'GRAPH_CLUSTERING', 'GRAPH_MATCHING', 'LINK_PREDICTION',
-            # 'REGRESSION', 'TIME_SERIES_FORECASTING', 'VERTEX_NOMINATION'
-            "inputType": "table",  # See SEMANTIC_TYPES.keys() for range of values
-            "output": "model_step",  # Name of the final step generating the prediction
-            "target": "extract_target_step",  # Name of the step generating the ground truth
-            "steps": TemplateSteps.dsbox_generic_steps() + [
-                {
-                    "name": "model_step",
-                    "runtime": {
-                        "cross_validation": 10,
-                        "stratified": True
-                    },
-                    "primitives": [
-                        {
-                            "primitive":
-                                "d3m.primitives.sklearn_wrap.SKExtraTreesClassifier",
-                            "hyperparameters":
-                                {
-                                    'bootstrap': [True, False],
-                                    'max_depth': [15, 30, None],
-                                    'min_samples_leaf': [1, 2, 4],
-                                    'min_samples_split': [2, 5, 10],
-                                    'max_features': ['auto', 'sqrt'],
-                                    'n_estimators': [10, 50, 100]
-                                }
-                        },
-                    ],
-                    "inputs": ["data", "target"]
-                }
-            ]
-        }
-
-    # @override
-    def importance(datset, problem_description):
-        return 7
-
-
-class GradientBoostingClassificationTemplate(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "gradient_boosting_classification_template",
-            "taskSubtype": {TaskSubtype.BINARY.name, TaskSubtype.MULTICLASS.name},
-            "taskType": TaskType.CLASSIFICATION.name,
-            # See TaskType, range include 'CLASSIFICATION', 'CLUSTERING', 'COLLABORATIVE_FILTERING',
-            # 'COMMUNITY_DETECTION', 'GRAPH_CLUSTERING', 'GRAPH_MATCHING', 'LINK_PREDICTION',
-            # 'REGRESSION', 'TIME_SERIES_FORECASTING', 'VERTEX_NOMINATION'
-            "inputType": "table",  # See SEMANTIC_TYPES.keys() for range of values
-            "output": "model_step",  # Name of the final step generating the prediction
-            "target": "extract_target_step",  # Name of the step generating the ground truth
-            "steps": TemplateSteps.dsbox_generic_steps() + [
-                {
-                    "name": "model_step",
-                    "runtime": {
-                        "cross_validation": 10,
-                        "stratified": True
-                    },
-                    "primitives": [
-                        {
-                            "primitive":
-                                "d3m.primitives.sklearn_wrap.SKGradientBoostingClassifier",
-                            "hyperparameters":
-                                {
-                                    'max_depth': [2, 3, 5],
-                                    'n_estimators': [50, 75, 100],
-                                    'learning_rate': [0.1, 0.3, 0.5],
-                                    'min_samples_split': [2, 3],
-                                    'min_samples_leaf': [1, 2],
-                                }
-                        },
-                    ],
-                    "inputs": ["data", "target"]
-                }
-            ]
-        }
-
-    # @override
-    def importance(datset, problem_description):
-        return 7
-
-
-class SVCClassificationTemplate(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "svc_classification_template",
-            "taskSubtype": {TaskSubtype.BINARY.name, TaskSubtype.MULTICLASS.name},
-            "taskType": TaskType.CLASSIFICATION.name,
-            # See TaskType, range include 'CLASSIFICATION', 'CLUSTERING', 'COLLABORATIVE_FILTERING',
-            # 'COMMUNITY_DETECTION', 'GRAPH_CLUSTERING', 'GRAPH_MATCHING', 'LINK_PREDICTION',
-            # 'REGRESSION', 'TIME_SERIES_FORECASTING', 'VERTEX_NOMINATION'
-            "inputType": "table",  # See SEMANTIC_TYPES.keys() for range of values
-            "output": "model_step",  # Name of the final step generating the prediction
-            "target": "extract_target_step",  # Name of the step generating the ground truth
-            "steps": TemplateSteps.dsbox_generic_steps() + [
-                {
-                    "name": "model_step",
-                    "runtime": {
-                        "cross_validation": 10,
-                        "stratified": True
-                    },
-                    "primitives": [
-                        {
-                            "primitive":
-                                "d3m.primitives.sklearn_wrap.SKSVC",
-                            "hyperparameters":
-                                {
-                                    'C': [0.8, 1.0, 1.2],
-                                    'kernel': ['rbf', 'poly'],
-                                    'degree': [2, 3, 4],
-                                }
-                        },
-                    ],
-                    "inputs": ["data", "target"]
-                }
-            ]
-        }
-
-    # @override
-    def importance(datset, problem_description):
-        return 7
-
-
-class ClassificationWithSelection(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "classification_with_feature_selection",
-            "taskSubtype": {TaskSubtype.BINARY.name, TaskSubtype.MULTICLASS.name},
-            "taskType": TaskType.CLASSIFICATION.name,
-            "inputType": "table",  # See SEMANTIC_TYPES.keys() for range of values
-            "output": "model_step",  # Name of the final step generating the prediction
-            "target": "extract_target_step",  # Name of the step generating the ground truth
-            "steps": TemplateSteps.human_steps() + TemplateSteps.dsbox_feature_selector("classification") +
-                [
-                    {
-                        "name":"model_step",
-                        "primitives":[
-                            {
-                                "primitive":"d3m.primitives.sklearn_wrap.SKSGDClassifier",
-                                "hyperparameters":{
-                                    "loss": ['log', 'hinge', 'squared_hinge', 'perceptron'],
-                                    "alpha": [float(x) for x in np.logspace(-6, -1.004, 7)],
-                                    "l1_ratio": [float(x) for x in np.logspace(-9, -0.004, 7)],
-                                    "penalty": ['elasticnet', 'l2']
-                                }
-                            },
-                            {
-                                "primitive":
-                                    "d3m.primitives.sklearn_wrap.SKGradientBoostingClassifier",
-                                "hyperparameters":
-                                    {
-                                        'max_depth': [2, 5],
-                                        'n_estimators': [50, 100],
-                                        'learning_rate': [0.1, 0.3],
-                                        'min_samples_split': [2, 3],
-                                        'min_samples_leaf': [1, 2],
-                                    }
-                            },
-
-                        ],
-                        "inputs":["feature_selector_step","extract_target_step"]
-                    }
-                ]
-        }
-
-    # @override
-    def importance(datset, problem_description):
-        return 7
-
-
-class dsboxClassificationTemplate(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "dsbox_classification_template",
-            "taskSubtype": {TaskSubtype.BINARY.name, TaskSubtype.MULTICLASS.name},
-            "taskType": TaskType.CLASSIFICATION.name,
-            # See TaskType, range include 'CLASSIFICATION', 'CLUSTERING',
-            # 'COLLABORATIVE_FILTERING', 'COMMUNITY_DETECTION', 'GRAPH_CLUSTERING',
-            # 'GRAPH_MATCHING', 'LINK_PREDICTION', 'REGRESSION', 'TIME_SERIES_FORECASTING',
-            # 'VERTEX_NOMINATION'
-            "inputType": "table",  # See SEMANTIC_TYPES.keys() for range of values
-            "output": "model_step",  # Name of the final step generating the prediction
-            "target": "extract_target_step",  # Name of the step generating the ground truth
-            "steps": [
-                *TemplateSteps.dsbox_preprocessing(
-                    clean_name="clean_step",
-                    target_name="extract_target_step"
-                ),
-                *TemplateSteps.dsbox_encoding(clean_name="clean_step",
-                                encoded_name="encoder_step"),
-
-                *TemplateSteps.dsbox_imputer(encoded_name="encoder_step",
-                               impute_name="impute_step"),
-
-                # *dimensionality_reduction(feature_name="impute_step",
-                #                           dim_reduce_name="dim_reduce_step"),
-
-
-
-                *TemplateSteps.classifier_model(feature_name="impute_step",
-                                  target_name='extract_target_step'),
-            ]
-        }
-
-    # @override
-    def importance(datset, problem_description):
-        return 7
-
-
-################################################################################################################
-####################################   General Regression Templates    #########################################
-################################################################################################################
-
-################################################################################################
-# A regression template encompassing several algorithms
-################################################################################################
-class DefaultRegressionTemplate(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "default_regression_template",
-            "taskSubtype": {TaskSubtype.UNIVARIATE.name, TaskSubtype.MULTIVARIATE.name},
-            "taskType": TaskType.REGRESSION.name,
-            # See TaskType, range include 'CLASSIFICATION', 'CLUSTERING', 'COLLABORATIVE_FILTERING',
-            # 'COMMUNITY_DETECTION', 'GRAPH_CLUSTERING', 'GRAPH_MATCHING', 'LINK_PREDICTION',
-            # 'REGRESSION', 'TIME_SERIES_FORECASTING', 'VERTEX_NOMINATION'
-            "inputType": "table",  # See SEMANTIC_TYPES.keys() for range of values
-            "output": "model_step",  # Name of the final step generating the prediction
-            "target": "extract_target_step",  # Name of the step generating the ground truth
-            "steps": TemplateSteps.dsbox_generic_steps() + [
-                {
-                    "name": "model_step",
-                    "runtime": {
-                        "cross_validation": 10,
-                        "stratified": False
-                    },
-                    "primitives": [
-                        {
-                            "primitive":
-                                "d3m.primitives.sklearn_wrap.SKGradientBoostingRegressor",
-                            "hyperparameters":
-                                {
-                                    'max_depth': [2, 3, 4, 5],
-                                    'n_estimators': [100, 130, 165, 200],
-                                    'learning_rate': [0.1, 0.23, 0.34, 0.5],
-                                    'min_samples_split': [2, 3],
-                                    'min_samples_leaf': [1, 2],
-                                }
-                        },
-                        {
-                            "primitive":
-                                "d3m.primitives.sklearn_wrap.SKExtraTreesRegressor",
-                            "hyperparameters":
-                                {
-                                    'bootstrap': [True, False],
-                                    'max_depth': [15, 30, None],
-                                    'min_samples_leaf': [1, 2, 4],
-                                    'min_samples_split': [2, 5, 10],
-                                    'max_features': ['auto', 'sqrt'],
-                                    'n_estimators': [10, 50, 100]
-                                }
-                        },
-                        {
-                            "primitive":
-                                "d3m.primitives.sklearn_wrap.SKRandomForestRegressor",
-                            "hyperparameters":
-                                {
-                                    'bootstrap': [True, False],
-                                    'max_depth': [15, 30, None],
-                                    'min_samples_leaf': [1, 2, 4],
-                                    'min_samples_split': [2, 5, 10],
-                                    'max_features': ['auto', 'sqrt'],
-                                    'n_estimators': [10, 50, 100]
-                                }
-                        },
-                    ],
-                    "inputs": ["data", "target"]
-                }
-            ]
-        }
-
-    # @override
-    def importance(datset, problem_description):
-        return 7
-
-
-class SVRRegressionTemplate(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "svr_regression_template",
-            "taskSubtype": {TaskSubtype.UNIVARIATE.name, TaskSubtype.MULTIVARIATE.name},
-            "taskType": TaskType.REGRESSION.name,
-            # See TaskType, range include 'CLASSIFICATION', 'CLUSTERING', 'COLLABORATIVE_FILTERING',
-            # 'COMMUNITY_DETECTION', 'GRAPH_CLUSTERING', 'GRAPH_MATCHING', 'LINK_PREDICTION',
-            # 'REGRESSION', 'TIME_SERIES_FORECASTING', 'VERTEX_NOMINATION'
-            "inputType": "table",  # See SEMANTIC_TYPES.keys() for range of values
-            "output": "model_step",  # Name of the final step generating the prediction
-            "target": "extract_target_step",  # Name of the step generating the ground truth
-            "steps": TemplateSteps.dsbox_generic_steps() + [
-                {
-                    "name": "model_step",
-                    "runtime": {
-                        "cross_validation": 10,
-                        "stratified": True
-                    },
-                    "primitives": [
-                        {
-                            "primitive":
-                                "d3m.primitives.sklearn_wrap.SKSVR",
-                            "hyperparameters":
-                                {
-                                    'C': [0.8, 1.0, 1.2],
-                                    'kernel': ['rbf', 'poly'],
-                                    'degree': [2, 3, 4, 5],
-                                }
-                        },
-                    ],
-                    "inputs": ["data", "target"]
-                }
-            ]
-        }
-
-    # @override
-    def importance(datset, problem_description):
-        return 7
-
-
-class GradientBoostingRegressionTemplate(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "gradient_boosting_regression_template",
-            "taskSubtype": {TaskSubtype.UNIVARIATE.name, TaskSubtype.MULTIVARIATE.name},
-            "taskType": TaskType.REGRESSION.name,
-            # See TaskType, range include 'CLASSIFICATION', 'CLUSTERING', 'COLLABORATIVE_FILTERING',
-            # 'COMMUNITY_DETECTION', 'GRAPH_CLUSTERING', 'GRAPH_MATCHING', 'LINK_PREDICTION',
-            # 'REGRESSION', 'TIME_SERIES_FORECASTING', 'VERTEX_NOMINATION'
-            "inputType": "table",  # See SEMANTIC_TYPES.keys() for range of values
-            "output": "model_step",  # Name of the final step generating the prediction
-            "target": "extract_target_step",  # Name of the step generating the ground truth
-            "steps": TemplateSteps.dsbox_generic_steps() + [
-                {
-                    "name": "model_step",
-                    "runtime": {
-                        "cross_validation": 10,
-                    },
-                    "primitives": [
-                        {
-                            "primitive":
-                                "d3m.primitives.sklearn_wrap.SKGradientBoostingRegressor",
-                            "hyperparameters":
-                                {
-                                    'max_depth': [2, 3, 5],
-                                    'n_estimators': [100, 150, 200],
-                                    'learning_rate': [0.1, 0.3, 0.5],
-                                    'min_samples_split': [2, 3],
-                                    'min_samples_leaf': [1, 2],
-                                }
-                        },
-                    ],
-                    "inputs": ["data", "target"]
-                }
-            ]
-        }
-
-    # @override
-    def importance(datset, problem_description):
-        return 7
-
-
-class ExtraTreesRegressionTemplate(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "extra_trees_regression_template",
-            "taskSubtype": {TaskSubtype.UNIVARIATE.name, TaskSubtype.MULTIVARIATE.name},
-            "taskType": TaskType.REGRESSION.name,
-            # See TaskType, range include 'CLASSIFICATION', 'CLUSTERING', 'COLLABORATIVE_FILTERING',
-            # 'COMMUNITY_DETECTION', 'GRAPH_CLUSTERING', 'GRAPH_MATCHING', 'LINK_PREDICTION',
-            # 'REGRESSION', 'TIME_SERIES_FORECASTING', 'VERTEX_NOMINATION'
-            "inputType": "table",  # See SEMANTIC_TYPES.keys() for range of values
-            "output": "model_step",  # Name of the final step generating the prediction
-            "target": "extract_target_step",  # Name of the step generating the ground truth
-            "steps": TemplateSteps.dsbox_generic_steps() + [
-                {
-                    "name": "model_step",
-                    "runtime": {
-                        "cross_validation": 10,
-                    },
-                    "primitives": [
-                        {
-                            "primitive":
-                                "d3m.primitives.sklearn_wrap.SKExtraTreesRegressor",
-                            "hyperparameters":
-                                {
-                                    'bootstrap': [True, False],
-                                    'max_depth': [15, 30, None],
-                                    'min_samples_leaf': [1, 2, 4],
-                                    'min_samples_split': [2, 5, 10],
-                                    'max_features': ['auto', 'sqrt'],
-                                    'n_estimators': [10, 50, 100]
-                                }
-                        },
-                    ],
-                    "inputs": ["data", "target"]
-                }
-            ]
-        }
-
-    # @override
-    def importance(datset, problem_description):
-        return 7
-
-
-class RandomForestRegressionTemplate(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "random_forest_regression_template",
-            "taskSubtype": {TaskSubtype.UNIVARIATE.name, TaskSubtype.MULTIVARIATE.name},
-            "taskType": TaskType.REGRESSION.name,
-            # See TaskType, range include 'CLASSIFICATION', 'CLUSTERING', 'COLLABORATIVE_FILTERING',
-            # 'COMMUNITY_DETECTION', 'GRAPH_CLUSTERING', 'GRAPH_MATCHING', 'LINK_PREDICTION',
-            # 'REGRESSION', 'TIME_SERIES_FORECASTING', 'VERTEX_NOMINATION'
-            "inputType": "table",  # See SEMANTIC_TYPES.keys() for range of values
-            "output": "model_step",  # Name of the final step generating the prediction
-            "target": "extract_target_step",  # Name of the step generating the ground truth
-            "steps": TemplateSteps.dsbox_generic_steps() + [
-                {
-                    "name": "model_step",
-                    "runtime": {
-                        "cross_validation": 10,
-                    },
-                    "primitives": [
-                        {
-                            "primitive":
-                                "d3m.primitives.sklearn_wrap.SKRandomForestRegressor",
-                            "hyperparameters":
-                                {
-                                    'bootstrap': [True, False],
-                                    'max_depth': [15, 30, None],
-                                    'min_samples_leaf': [1, 2, 4],
-                                    'min_samples_split': [2, 5, 10],
-                                    'max_features': ['auto', 'sqrt'],
-                                    'n_estimators': [10, 50, 100]
-                                }
-                        },
-                    ],
-                    "inputs": ["data", "target"]
-                }
-            ]
-        }
-
-    # @override
-    def importance(datset, problem_description):
-        return 7
-
-
-class RegressionWithSelection(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "regression_with_feature_selection",
-            "taskSubtype": {TaskSubtype.UNIVARIATE.name, TaskSubtype.MULTIVARIATE.name},
-            "taskType": TaskType.REGRESSION.name,
-            "inputType": "table",  # See SEMANTIC_TYPES.keys() for range of values
-            "output": "model_step",  # Name of the final step generating the prediction
-            "target": "extract_target_step",  # Name of the step generating the ground truth
-            "steps": TemplateSteps.human_steps() + TemplateSteps.dsbox_feature_selector("regression") +
-                [
-                    {
-                        "name":"model_step",
-                        "primitives":[
-                            {
-                                "primitive":"d3m.primitives.sklearn_wrap.SKSGDRegressor",
-                                "hyperparameters":{
-                                    "loss":['squared_loss', 'huber'],
-                                    "alpha":[float(x) for x in np.logspace(-5, -1.004, 7)],#cannot reach 0.1
-                                    "l1_ratio":[0.01,0.15, 0.3, 0.5, 0.6, 0.7, 0.9], #cannot reach 1
-                                    "learning_rate": ['optimal', 'invscaling']
-                                }
-                            },
-                            {
-                                "primitive":
-                                    "d3m.primitives.sklearn_wrap.SKGradientBoostingRegressor",
-                                "hyperparameters":
-                                    {
-                                        'max_depth': [2, 3, 5],
-                                        'n_estimators': [100, 150, 200],
-                                        'learning_rate': [0.1, 0.3, 0.5],
-                                        'min_samples_split': [2, 3],
-                                        'min_samples_leaf': [1, 2],
-                                    }
-                            },
-                        ],
-                        "inputs":["feature_selector_step","extract_target_step"]
-                    }
-                ]
-        }
-
-    # @override
-    def importance(datset, problem_description):
-        return 7
-
-
-class dsboxRegressionTemplate(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "dsbox_regression_template",
-            "taskType": TaskType.REGRESSION.name,
-            "taskSubtype": {TaskSubtype.UNIVARIATE.name, TaskSubtype.MULTIVARIATE.name},
-            # See TaskType, range include 'CLASSIFICATION', 'CLUSTERING', 'COLLABORATIVE_FILTERING',
-            # 'COMMUNITY_DETECTION', 'GRAPH_CLUSTERING', 'GRAPH_MATCHING', 'LINK_PREDICTION',
-            # 'REGRESSION', 'TIME_SERIES_FORECASTING', 'VERTEX_NOMINATION'
-            "inputType": "table",  # See SEMANTIC_TYPES.keys() for range of values
-            "output": "model_step",  # Name of the final step generating the prediction
-            "target": "extract_target_step",  # Name of the step generating the ground truth
-            "steps": [
-                *TemplateSteps.dsbox_preprocessing(
-                    clean_name="clean_step",
-                    target_name="extract_target_step"
-                ),
-                *TemplateSteps.dsbox_encoding(
-                    clean_name="clean_step",
-                    encoded_name="encoder_step"
-                ),
-                *TemplateSteps.dsbox_imputer(
-                    encoded_name="encoder_step",
-                    impute_name="impute_step"
-                ),
-                *TemplateSteps.regression_model(
-                    feature_name="impute_step",
-                    target_name="extract_target_step"
-                ),
-            ]
-        }
-
-    # @override
-    def importance(datset, problem_description):
-        return 7
-
-
-
-################################################################################################################
-#####################################   Templates for Large Datasets  ##########################################
-################################################################################################################
-
-
-class Large_column_number_with_numerical_only_classification(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "Large_column_number_with_numerical_only_classification",
-            "taskType": TaskType.CLASSIFICATION.name,
-            # See TaskType, range include 'CLASSIFICATION', 'CLUSTERING', 'COLLABORATIVE_FILTERING',
-            # 'COMMUNITY_DETECTION', 'GRAPH_CLUSTERING', 'GRAPH_MATCHING', 'LINK_PREDICTION',
-            # 'REGRESSION', 'TIME_SERIES_FORECASTING', 'VERTEX_NOMINATION'
-            "taskSubtype": {TaskSubtype.MULTICLASS.name, TaskSubtype.BINARY.name},
-            "inputType": {"table","large_column_number"},  # See SEMANTIC_TYPES.keys() for range of values
-            "output": "model_step",  # Name of the final step generating the prediction
-            "target": "extract_target_step",  # Name of the step generating the ground truth
-            "steps": [
-                {
-                    "name": "denormalize_step",
-                    "primitives": ["d3m.primitives.dsbox.Denormalize"],
-                    "inputs": ["template_input"]
-                },
-                {
-                    "name": "to_dataframe_step",
-                    "primitives": ["d3m.primitives.datasets.DatasetToDataFrame"],
-                    "inputs": ["denormalize_step"]
-                },
-                {
-                    "name": "extract_attribute_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
-                        "hyperparameters":
-                            {
-                                'semantic_types': ('https://metadata.datadrivendiscovery.org/types/Attribute',),
-                                'use_columns': (),
-                                'exclude_columns': ()
-                            }
-                    }],
-                    "inputs": ["to_dataframe_step"]
-                },
-                {
-                    "name": "extract_target_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
-                        "hyperparameters":
-                            {'semantic_types': ('https://metadata.datadrivendiscovery.org/types/Target',
-                                                'https://metadata.datadrivendiscovery.org/types/SuggestedTarget',),
-                             'use_columns': (),
-                             'exclude_columns': ()
-                             }
-                    }],
-                    "inputs": ["to_dataframe_step"]
-                },
-                # {
-                #     "name": "encode1_step",
-                #     # "primitives": ["d3m.primitives.dsbox.UnaryEncoder","d3m.primitives.dsbox.DoNothing"],
-                #     "primitives": ["d3m.primitives.dsbox.UnaryEncoder"],
-                #     "inputs": ["extract_attribute_step"]
-                # },
-                {
-                    "name": "encode2_step",
-                    "primitives": ["d3m.primitives.dsbox.Labler"],
-                    "inputs": ["extract_attribute_step"]
-                },
-                # {
-                #     "name": "impute_step",
-                #     "primitives": ["d3m.primitives.dsbox.MeanImputation"],
-                #     "inputs": ["encode2_step"]
-                # },
-                {
-                    "name": "cast_1_step",
-                    "primitives": [
-                        "d3m.primitives.dsbox.DoNothing",
-                        {
-                            "primitive": "d3m.primitives.data.CastToType",
-                            "hyperparameters": {"type_to_cast": ["float"]}
-                        },
-
-                    ],
-                    "inputs": ["encode2_step"]
-                },
-                {
-                    "name": "model_step",
-                    "primitives": ["d3m.primitives.sklearn_wrap.SKRandomForestClassifier"],
-                    "inputs": ["cast_1_step", "extract_target_step"]
-                },
-            ]
-        }
-
-    # @override
-    def importance(datset, problem_description):
-        return 7
-
-class Large_column_number_with_numerical_only_regression(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "Large_column_number_with_numerical_only_regression",
-            "taskSubtype": {TaskSubtype.UNIVARIATE.name, TaskSubtype.MULTIVARIATE.name},
-            "taskType": TaskType.REGRESSION.name,
-            "inputType": {"table","large_column_number"},  # See SEMANTIC_TYPES.keys() for range of values
-            "output": "model_step",  # Name of the final step generating the prediction
-            "target": "extract_target_step",  # Name of the step generating the ground truth
-            "steps": [
-                {
-                    "name": "denormalize_step",
-                    "primitives": ["d3m.primitives.dsbox.Denormalize"],
-                    "inputs": ["template_input"]
-                },
-                {
-                    "name": "to_dataframe_step",
-                    "primitives": ["d3m.primitives.datasets.DatasetToDataFrame"],
-                    "inputs": ["denormalize_step"]
-                },
-                {
-                    "name": "extract_attribute_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
-                        "hyperparameters":
-                            {
-                                'semantic_types': ('https://metadata.datadrivendiscovery.org/types/Attribute',),
-                                'use_columns': (),
-                                'exclude_columns': ()
-                            }
-                    }],
-                    "inputs": ["to_dataframe_step"]
-                },
-                {
-                    "name": "extract_target_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
-                        "hyperparameters":
-                            {'semantic_types': ('https://metadata.datadrivendiscovery.org/types/Target',
-                                                'https://metadata.datadrivendiscovery.org/types/SuggestedTarget',),
-                             'use_columns': (),
-                             'exclude_columns': ()
-                             }
-                    }],
-                    "inputs": ["to_dataframe_step"]
-                },
-                # {
-                #     "name": "encode1_step",
-                #     # "primitives": ["d3m.primitives.dsbox.UnaryEncoder","d3m.primitives.dsbox.DoNothing"],
-                #     "primitives": ["d3m.primitives.dsbox.UnaryEncoder"],
-                #     "inputs": ["extract_attribute_step"]
-                # },
-                {
-                    "name": "encode2_step",
-                    "primitives": ["d3m.primitives.dsbox.Labler"],
-                    "inputs": ["extract_attribute_step"]
-                },
-                # {
-                #     "name": "impute_step",
-                #     "primitives": ["d3m.primitives.dsbox.MeanImputation"],
-                #     "inputs": ["encode2_step"]
-                # },
-                {
-                    "name": "cast_1_step",
-                    "primitives": [
-                        "d3m.primitives.dsbox.DoNothing",
-                        {
-                            "primitive": "d3m.primitives.data.CastToType",
-                            "hyperparameters": {"type_to_cast": ["float"]}
-                        },
-
-                    ],
-                    "inputs": ["encode2_step"]
-                },
-                {
-                    "name": "model_step",
-                    "primitives": ["d3m.primitives.sklearn_wrap.SKRandomForestRegressor"],
-                    "inputs": ["cast_1_step", "extract_target_step"]
-                },
-            ]
-        }
-
-    # @override
-    def importance(datset, problem_description):
-        return 7
-
-
-################################################################################################################
-#####################################   TimeSeriesForcasting Templates  ########################################
-################################################################################################################
-
-
-class DefaultTimeSeriesForcastingTemplate(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "Default_Time_Series_Forcasting_Template",
-            "taskType": TaskType.TIME_SERIES_FORECASTING.name,
-            # See TaskType, range include 'CLASSIFICATION', 'CLUSTERING', 'COLLABORATIVE_FILTERING',
-            # 'COMMUNITY_DETECTION', 'GRAPH_CLUSTERING', 'GRAPH_MATCHING', 'LINK_PREDICTION',
-            # 'REGRESSION', 'TIME_SERIES_FORECASTING', 'VERTEX_NOMINATION'
-            "taskSubtype": "NONE",
-            "inputType": {"table", "timeseries"},  # See SEMANTIC_TYPES.keys() for range of values
-            "output": "model_step",  # Name of the final step generating the prediction
-            "target": "extract_target_step",  # Name of the step generating the ground truth
-            "steps": [
-
-                {
-                    "name": "denormalize_step",
-                    "primitives": ["d3m.primitives.dsbox.Denormalize"],
-                    "inputs": ["template_input"]
-                },
-                {
-                    "name": "to_dataframe_step",
-                    "primitives": ["d3m.primitives.datasets.DatasetToDataFrame"],
-                    "inputs": ["denormalize_step"]
-                },
-                # read Y value
-                {
-                    "name": "extract_target_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
-                        "hyperparameters":
-                            {
-                                'semantic_types': ('https://metadata.datadrivendiscovery.org/types/TrueTarget',),
-                                'use_columns': (),
-                                'exclude_columns': ()
-                            }
-                        }],
-                    "inputs": ["to_dataframe_step"]
-                },
-
-                # read X value
-                {
-                    "name": "extract_attribute_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
-                        "hyperparameters":
-                            {
-                                'semantic_types': (
-                                    'https://metadata.datadrivendiscovery.org/types/Attribute',),
-                                'use_columns': (),
-                                'exclude_columns': ()
-                            }
-                    }],
-                    "inputs": ["to_dataframe_step"]
-                },
-                {
-                    "name": "profile_step",
-                    "primitives": ["d3m.primitives.dsbox.Profiler"],
-                    "inputs": ["extract_attribute_step"]
-                },
-                {
-                    "name": "clean_step",
-                    "primitives": [
-                        "d3m.primitives.dsbox.CleaningFeaturizer",
-                        "d3m.primitives.dsbox.DoNothing",
-                    ],
-                    "inputs": ["profile_step"]
-                },
-                {
-                    "name": "timeseries_to_list_step",
-                    "primitives": ["d3m.primitives.dsbox.TimeseriesToList"],
-                    "inputs": ["clean_step"]
-                },
-
-                {
-                    "name": "random_projection_step",
-                    "primitives": ["d3m.primitives.dsbox.RandomProjectionTimeSeriesFeaturization"],
-                    "inputs": ["timeseries_to_list_step"]
-                },
-                {
-                    "name": "cast_1_step",
-                    "primitives": [
-                        {
-                            "primitive": "d3m.primitives.data.CastToType",
-                            "hyperparameters": {"type_to_cast": ["float"]}
-                        },
-                        {
-                            "primitive": "d3m.primitives.dsbox.DoNothing",
-                            "hyperparameters": {}
-                        }
-                    ],
-                    "inputs": ["extract_target_step"]
-                },
-                {
-                    "name": "random_forest_step",
-                    "primitives": ["d3m.primitives.sklearn_wrap.SKRandomForestRegressor"],
-                    "inputs": ["random_projection_step", "cast_1_step"]
-                },
-            ]
-        }
-
-    # @override
-    def importance(datset, problem_description):
-        return 7
-
-
-class TimeSeriesForcastingTestingTemplate(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "TimeSeries_Forcasting_Testing_emplate",
-            "taskType": TaskType.TIME_SERIES_FORECASTING.name,
-            # See TaskType, range include 'CLASSIFICATION', 'CLUSTERING', 'COLLABORATIVE_FILTERING',
-            # 'COMMUNITY_DETECTION', 'GRAPH_CLUSTERING', 'GRAPH_MATCHING', 'LINK_PREDICTION',
-            # 'REGRESSION', 'TIME_SERIES_FORECASTING', 'VERTEX_NOMINATION'
-            "taskSubtype": "NONE",
-            "inputType": {"table", "timeseries"},  # See SEMANTIC_TYPES.keys() for range of values
-            "output": "model_step",  # Name of the final step generating the prediction
-            "target": "extract_target_step",  # Name of the step generating the ground truth
-            "steps": [
-                {
-                    "name": "denormalize_step",
-                    "primitives": ["d3m.primitives.dsbox.Denormalize"],
-                    "inputs": ["template_input"]
-                },
-                {
-                    "name": "to_dataframe_step",
-                    "primitives": ["d3m.primitives.datasets.DatasetToDataFrame"],
-                    "inputs": ["denormalize_step"]
-                },
-                {
-                    "name": "extract_attribute_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
-                        "hyperparameters":
-                            {
-                                'semantic_types': (
-                                    'https://metadata.datadrivendiscovery.org/types/Attribute',),
-                                'use_columns': (),
-                                'exclude_columns': ()
-                            }
-                    }],
-                    "inputs": ["to_dataframe_step"]
-                },
-                {
-                    "name":"profiler_step",
-                    "primitives": ["d3m.primitives.dsbox.Profiler"],
-                    "inputs": ["extract_attribute_step"]
-                },
-                {
-                    "name":"data_clean_step",
-                    "primitives": ["d3m.primitives.dsbox.CleaningFeaturizer"],
-                    "inputs": ["profiler_step"]
-                },
-                {
-                    "name":"encoder_step",
-                    "primitives": ["d3m.primitives.dsbox.Encoder"],
-                    "inputs": ["data_clean_step"]
-                },
-                # read Y value
-                {
-                    "name": "extract_target_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
-                        "hyperparameters":
-                            {
-                                'semantic_types': (
-                                    'https://metadata.datadrivendiscovery.org/types/TrueTarget',),
-                                'use_columns': (),
-                                'exclude_columns': ()
-                            }
-                    }],
-                    "inputs": ["to_dataframe_step"]
-                },
-                {
-                    "name": "model_step",
-                    "primitives": ["d3m.primitives.sklearn_wrap.SKExtraTreesRegressor"],
-                    "inputs": ["encoder_step", "extract_target_step"]
-                },
-            ]
-        }
-
-    # @override
-    def importance(datset, problem_description):
-        return 7
-
-
-'''
-This template cannot run because of our templates' "input"/"output" schema
-'''
-class TimeSeriesForcastingTestingTemplate2(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "TimeSeries_Forcasting_Testing_emplate",
-            "taskType": TaskType.TIME_SERIES_FORECASTING.name,
-            # See TaskType, range include 'CLASSIFICATION', 'CLUSTERING', 'COLLABORATIVE_FILTERING',
-            # 'COMMUNITY_DETECTION', 'GRAPH_CLUSTERING', 'GRAPH_MATCHING', 'LINK_PREDICTION',
-            # 'REGRESSION', 'TIME_SERIES_FORECASTING', 'VERTEX_NOMINATION'
-            "taskSubtype": "NONE",
-            "inputType": {"table", "timeseries"},  # See SEMANTIC_TYPES.keys() for range of values
-            "output": "model_step",  # Name of the final step generating the prediction
-            "target": "extract_target_step",  # Name of the step generating the ground truth
-            "steps": [
-                {
-                    "name": "denormalize_step",
-                    "primitives": ["d3m.primitives.dsbox.Denormalize"],
-                    "inputs": ["template_input"]
-                },
-                {
-                    "name": "to_dataframe_step",
-                    "primitives": ["d3m.primitives.datasets.DatasetToDataFrame"],
-                    "inputs": ["denormalize_step"]
-                },
-                {
-                    "name": "extract_attribute_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
-                        "hyperparameters":
-                            {
-                                'semantic_types': (
-                                    'https://metadata.datadrivendiscovery.org/types/Attribute',),
-                                'use_columns': (),
-                                'exclude_columns': (3,)
-                            }
-                    }],
-                    "inputs": ["to_dataframe_step"]
-                },
-                {
-                    "name": "extract_time_series_file_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
-                        "hyperparameters":
-                            {
-                                'semantic_types': (
-                                    'https://metadata.datadrivendiscovery.org/types/FileName',),
-                                'use_columns': (),
-                                'exclude_columns': ()
-                            }
-                    }],
-                    "inputs": ["to_dataframe_step"]
-                },
-                {
-                    "name": "timeseries_to_list_step",
-                    "primitives": ["d3m.primitives.dsbox.TimeseriesToList"],
-                    "inputs": ["extract_time_series_file_step"]
-                },
-
-                {
-                    "name": "random_projection_step",
-                    "primitives": ["d3m.primitives.dsbox.RandomProjectionTimeSeriesFeaturization"],
-                    "inputs": ["timeseries_to_list_step"]
-                },
-                {
-                    "name":"profiler_step",
-                    "primitives": ["d3m.primitives.dsbox.Profiler"],
-                    "inputs": ["extract_attribute_step"]
-                },
-                {
-                    "name":"data_clean_step",
-                    "primitives": ["d3m.primitives.dsbox.CleaningFeaturizer"],
-                    "inputs": ["profiler_step"]
-                },
-                {
-                    "name":"encoder_step",
-                    "primitives": ["d3m.primitives.dsbox.Encoder"],
-                    "inputs": ["data_clean_step"]
-                },
-                {
-                    "name":"concat_step",
-                    "primitives":["d3m.primitives.data.HorizontalConcatPrimitive"],
-                    "inputs":["encoder_step","random_projection_step"]
-                },
-                # read Y value
-                {
-                    "name": "extract_target_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
-                        "hyperparameters":
-                            {
-                                'semantic_types': (
-                                    'https://metadata.datadrivendiscovery.org/types/TrueTarget',),
-                                'use_columns': (),
-                                'exclude_columns': ()
-                            }
-                    }],
-                    "inputs": ["to_dataframe_step"]
-                },
-                {
-                    "name": "model_step",
-                    "primitives": ["d3m.primitives.sklearn_wrap.SKExtraTreesRegressor"],
-                    "inputs": ["concat_step", "extract_target_step"]
-                },
-            ]
-        }
-
-    # @override
-    def importance(datset, problem_description):
-        return 7
-
-
-################################################################################################################
-#####################################   ObjectDetectionTemplates   #############################################
-################################################################################################################
-
-
-class TemporaryObjectDetectionTemplate(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "TemporaryObjectDetectionTemplate",
-            "taskType": TaskType.OBJECT_DETECTION.name,
-            # See TaskType, range include 'CLASSIFICATION', 'CLUSTERING', 'COLLABORATIVE_FILTERING',
-            # 'COMMUNITY_DETECTION', 'GRAPH_CLUSTERING', 'GRAPH_MATCHING', 'LINK_PREDICTION',
-            # 'REGRESSION', 'TIME_SERIES_FORECASTING', 'VERTEX_NOMINATION'
-            "taskSubtype": "NONE",
-            "inputType": {"table", "image"},  # See SEMANTIC_TYPES.keys() for range of values
-            "output": "model_step",  # Name of the final step generating the prediction
-            "target": "extract_target_step",  # Name of the step generating the ground truth
-            "steps": [
-                {
-                    "name": "denormalize_step",
-                    "primitives": ["d3m.primitives.dsbox.Denormalize"],
-                    "inputs": ["template_input"]
-                },
-                {
-                    "name": "to_dataframe_step",
-                    "primitives": ["d3m.primitives.datasets.DatasetToDataFrame"],
-                    "inputs": ["denormalize_step"]
-                },
-                # read X value
-                {
-                    "name": "extract_file_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
-                        "hyperparameters":
-                            {
-                                'semantic_types': (
-                                    'https://metadata.datadrivendiscovery.org/types/FileName',),
-                                'use_columns': (),
-                                'exclude_columns': ()
-                            }
-                    }],
-                    "inputs": ["to_dataframe_step"]
-                },
-                {
-                    "name": "to_tensor_step",
-                    "primitives": ["d3m.primitives.dsbox.DataFrameToTensor"],
-                    "inputs": ["extract_file_step"]
-                },
-                {
-                    "name": "image_processing_step",
-                    "primitives": ["d3m.primitives.dsbox.ResNet50ImageFeature"],
-                    "inputs": ["to_tensor_step"]
-                },
-                # read Y value
-                {
-                    "name": "extract_target_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
-                        "hyperparameters":
-                            {
-                                'semantic_types': (
-                                    'https://metadata.datadrivendiscovery.org/types/TrueTarget',),
-                                'use_columns': (),
-                                'exclude_columns': ()
-                            }
-                    }],
-                    "inputs": ["to_dataframe_step"]
-                },
-
-                {
-                    "name":"data_clean_step",
-                    "primitives": ["d3m.primitives.dsbox.CleaningFeaturizer"],
-                    "inputs": ["extract_target_step"]
-                },
-
-                {
-                    "name": "model_step",
-                    "primitives": ["d3m.primitives.sklearn_wrap.SKRandomForestRegressor"],
-                    "inputs": ["image_processing_step", "data_clean_step"]
-                },
-            ]
-        }
-
-    # @override
-    def importance(datset, problem_description):
-        return 7
-
-
-################################################################################################################
-#####################################   TimeSeriesProblemsTemplates   ##########################################
-################################################################################################################
-
-
-class DefaultTimeseriesCollectionTemplate(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "Default_timeseries_collection_template",
-            "taskType": TaskType.CLASSIFICATION.name,
-            # See TaskType, range include 'CLASSIFICATION', 'CLUSTERING', 'COLLABORATIVE_FILTERING',
-            # 'COMMUNITY_DETECTION', 'GRAPH_CLUSTERING', 'GRAPH_MATCHING', 'LINK_PREDICTION',
-            # 'REGRESSION', 'TIME_SERIES_FORECASTING', 'VERTEX_NOMINATION'
-            "taskSubtype": {TaskSubtype.BINARY.name, TaskSubtype.MULTICLASS.name},
-            "inputType": {"timeseries", "table"},  # See SEMANTIC_TYPES.keys() for range of values
-            "output": "random_forest_step",  # Name of the final step generating the prediction
-            "target": "extract_target_step",  # Name of the step generating the ground truth
-            "steps": [
-
-                {
-                    "name": "denormalize_step",
-                    "primitives": ["d3m.primitives.dsbox.Denormalize"],
-                    "inputs": ["template_input"]
-                },
-                {
-                    "name": "to_dataframe_step",
-                    "primitives": ["d3m.primitives.datasets.DatasetToDataFrame"],
-                    "inputs": ["denormalize_step"]
-                },
-                # read Y value
-                {
-                    "name": "extract_target_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
-                        "hyperparameters":
-                            {
-                                'semantic_types': ('https://metadata.datadrivendiscovery.org/types/TrueTarget',),
-                                'use_columns': (),
-                                'exclude_columns': ()
-                            }
-                        }],
-                    "inputs": ["to_dataframe_step"]
-                },
-                # {
-                #     "name": "column_parser_step",
-                #     "primitives": ["d3m.primitives.data.ColumnParser"],
-                #     "inputs": ["extract_target_step"]
-                # },
-
-                # read X value
-                {
-                    "name": "timeseries_to_list_step",
-                    "primitives": ["d3m.primitives.dsbox.TimeseriesToList"],
-                    "inputs": ["to_dataframe_step"]
-                },
-
-                {
-                    "name": "random_projection_step",
-                    "primitives": ["d3m.primitives.dsbox.RandomProjectionTimeSeriesFeaturization"],
-                    "inputs": ["timeseries_to_list_step"]
-                },
-
-                {
-                    "name": "random_forest_step",
-                    "primitives": ["d3m.primitives.sklearn_wrap.SKRandomForestClassifier"],
-                    "inputs": ["random_projection_step", "extract_target_step"]
-                },
-            ]
-        }
-
-    # @override
-    def importance(datset, problem_description):
-        return 7
-
-
-class DefaultTimeseriesRegressionTemplate(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "DefaultTimeseriesRegressionTemplate",
-            "taskSubtype": {TaskSubtype.UNIVARIATE.name, TaskSubtype.MULTIVARIATE.name},
-            "taskType": {TaskType.REGRESSION.name},
-            "inputType": {"timeseries", "table"},  # See SEMANTIC_TYPES.keys() for range of values
-            "output": "random_forest_step",  # Name of the final step generating the prediction
-            "target": "extract_target_step",  # Name of the step generating the ground truth
-            "steps": [
-
-                {
-                    "name": "denormalize_step",
-                    "primitives": ["d3m.primitives.dsbox.Denormalize"],
-                    "inputs": ["template_input"]
-                },
-                {
-                    "name": "to_dataframe_step",
-                    "primitives": ["d3m.primitives.datasets.DatasetToDataFrame"],
-                    "inputs": ["denormalize_step"]
-                },
-                # read Y value
-                {
-                    "name": "extract_target_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
-                        "hyperparameters":
-                            {
-                                'semantic_types': ('https://metadata.datadrivendiscovery.org/types/TrueTarget',),
-                                'use_columns': (),
-                                'exclude_columns': ()
-                            }
-                        }],
-                    "inputs": ["to_dataframe_step"]
-                },
-
-                # read X value
-                {
-                    "name": "extract_attribute_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
-                        "hyperparameters":
-                            {
-                                'semantic_types': (
-                                    'https://metadata.datadrivendiscovery.org/types/Attribute',),
-                                'use_columns': (),
-                                'exclude_columns': ()
-                            }
-                    }],
-                    "inputs": ["to_dataframe_step"]
-                },
-                {
-                    "name": "timeseries_to_list_step",
-                    "primitives": ["d3m.primitives.dsbox.TimeseriesToList"],
-                    "inputs": ["extract_attribute_step"]
-                },
-
-                {
-                    "name": "random_projection_step",
-                    "primitives": ["d3m.primitives.dsbox.RandomProjectionTimeSeriesFeaturization"],
-                    "inputs": ["timeseries_to_list_step"]
-                },
-                {
-                    "name": "cast_1_step",
-                    "primitives": [
-                        {
-                            "primitive": "d3m.primitives.data.CastToType",
-                            "hyperparameters": {"type_to_cast": ["float"]}
-                        },
-                        {
-                            "primitive": "d3m.primitives.dsbox.DoNothing",
-                            "hyperparameters": {}
-                        }
-                    ],
-                    "inputs": ["extract_target_step"]
-                },
-                {
-                    "name": "random_forest_step",
-                    "primitives": ["d3m.primitives.sklearn_wrap.SKRandomForestRegressor"],
-                    "inputs": ["random_projection_step", "cast_1_step"]
-                },
-            ]
-        }
-
-    # @override
-    def importance(datset, problem_description):
-        return 7
-
-
-'''
-This template never working because templates "input", "outputs" schema
-'''
-class UCHITimeSeriesClassificationTemplate(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "UCHI_Time_Series_Classification_Template",
-            "taskType": TaskType.CLASSIFICATION.name,
-            # See TaskType, range include 'CLASSIFICATION', 'CLUSTERING', 'COLLABORATIVE_FILTERING',
-            # 'COMMUNITY_DETECTION', 'GRAPH_CLUSTERING', 'GRAPH_MATCHING', 'LINK_PREDICTION',
-            # 'REGRESSION', 'TIME_SERIES_FORECASTING', 'VERTEX_NOMINATION'
-            "taskSubtype": {TaskSubtype.BINARY.name, TaskSubtype.MULTICLASS.name},
-            "inputType": "timeseries",  # See SEMANTIC_TYPES.keys() for range of values
-            "output": "model_step",  # Name of the final step generating the prediction
-            "target": "extract_target_step",  # Name of the step generating the ground truth
-            "steps": [
-                # {
-                #     "name": "denormalize_step",
-                #     "primitives": ["d3m.primitives.dsbox.Denormalize"],
-                #     "inputs": ["template_input"]
-                # },
-                # {
-                #     "name": "to_dataframe_step",
-                #     "primitives": ["d3m.primitives.datasets.DatasetToDataFrame"],
-                #     "inputs": ["denormalize_step"]
-                # },
-                # # read Y value
-                # {
-                #     "name": "extract_target_step",
-                #     "primitives": [{
-                #         "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
-                #         "hyperparameters":
-                #             {
-                #                 'semantic_types': (
-                #                 'https://metadata.datadrivendiscovery.org/types/Target',
-                #                 'https://metadata.datadrivendiscovery.org/types/SuggestedTarget',),
-                #                 'use_columns': (),
-                #                 'exclude_columns': ()
-                #             }
-                #     }],
-                #     "inputs": ["to_dataframe_step"]
-                # },
-
-                {
-                    "name": "model_step",
-                    "primitives": ["d3m.primitives.datasmash.d3m_XG2"],
-                    "inputs": ["template_input", "template_input"]
-                }
-            ]
-        }
-
-    # @override
-    def importance(datset, problem_description):
-        return 7
-
-
-################################################################################################################
-#####################################   ImageProblemsTemplates   ###############################################
-################################################################################################################
-
-
-class TA1VggImageProcessingRegressionTemplate(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "TA1VggImageProcessingRegressionTemplate",
-            "taskType": TaskType.REGRESSION.name,
-            # See TaskType, range include 'CLASSIFICATION', 'CLUSTERING', 'COLLABORATIVE_FILTERING', 'COMMUNITY_DETECTION', 'GRAPH_CLUSTERING', 'GRAPH_MATCHING', 'LINK_PREDICTION', 'REGRESSION', 'TIME_SERIES_FORECASTING', 'VERTEX_NOMINATION'
-            "taskSubtype": {TaskSubtype.UNIVARIATE.name, TaskSubtype.MULTIVARIATE.name},
-            "inputType": "image",  # See SEMANTIC_TYPES.keys() for range of values
-            "output": "regressor_step",  # Name of the final step generating the prediction
-            "target": "extract_target_step",  # Name of the step generating the ground truth
-            "steps": [
-                {
-                    "name": "denormalize_step",
-                    "primitives": ["d3m.primitives.dsbox.Denormalize"],
-                    "inputs": ["template_input"]
-                },
-                {
-                    "name": "to_dataframe_step",
-                    "primitives": ["d3m.primitives.datasets.DatasetToDataFrame"],
-                    "inputs": ["denormalize_step"]
-                },
-                # read Y value
-                {
-                    "name": "extract_target_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
-                        "hyperparameters":
-                            {
-                                'semantic_types': ('https://metadata.datadrivendiscovery.org/types/TrueTarget',),
-                                'use_columns': (),
-                                'exclude_columns': ()
-                            }
-                        }],
-                    "inputs": ["to_dataframe_step"]
-                },
-                # {
-                #     "name": "column_parser_step",
-                #     "primitives": ["d3m.primitives.data.ColumnParser"],
-                #     "inputs": ["to_dataframe_step"]
-                # },
-                # read X value
-                {
-                    "name": "dataframe_to_tensor",
-                    "primitives": ["d3m.primitives.dsbox.DataFrameToTensor"],
-                    "inputs": ["to_dataframe_step"]
-                },
-                {
-                    "name": "feature_extraction",
-                    "primitives": ["d3m.primitives.dsbox.ResNet50ImageFeature"],
-                    # or "primitives": ["d3m.primitives.dsbox.Vgg16ImageFeature"],
-                    "inputs": ["dataframe_to_tensor"]
-                },
-                {
-                    "name": "PCA_step",
-                    "primitives": ["d3m.primitives.sklearn_wrap.SKPCA"],
-                    "inputs": ["feature_extraction"]
-                },
-
-                {
-                    "name": "regressor_step",
-                    "primitives": ["d3m.primitives.sklearn_wrap.SKRandomForestRegressor"],
-                    "inputs": ["PCA_step", "extract_target_step"]
-                },
-            ]
-        }
-
-    def importance(datset, problem_description):
-        return 7
-
-
-class TA1DefaultImageProcessingRegressionTemplate(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "Default_image_processing_regression_template",
-            "taskType": TaskType.REGRESSION.name,
-            # See TaskType, range include 'CLASSIFICATION', 'CLUSTERING', 'COLLABORATIVE_FILTERING',
-            # 'COMMUNITY_DETECTION', 'GRAPH_CLUSTERING', 'GRAPH_MATCHING', 'LINK_PREDICTION',
-            # 'REGRESSION', 'TIME_SERIES_FORECASTING', 'VERTEX_NOMINATION'
-            "taskSubtype": {TaskSubtype.UNIVARIATE.name, TaskSubtype.MULTIVARIATE.name},
-            "inputType": "image",  # See SEMANTIC_TYPES.keys() for range of values
-            "output": "regressor_step",  # Name of the final step generating the prediction
-            "target": "extract_target_step",  # Name of the step generating the ground truth
-            "steps": [
-
-                {
-                    "name": "denormalize_step",
-                    "primitives": ["d3m.primitives.dsbox.Denormalize"],
-                    "inputs": ["template_input"]
-                },
-                {
-                    "name": "to_dataframe_step",
-                    "primitives": ["d3m.primitives.datasets.DatasetToDataFrame"],
-                    "inputs": ["denormalize_step"]
-                },
-                # read Y value
-                {
-                    "name": "extract_target_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
-                        "hyperparameters":
-                            {
-                                'semantic_types': ('https://metadata.datadrivendiscovery.org/types/TrueTarget',),
-                                'use_columns': (),
-                                'exclude_columns': ()
-                            }
-                        }],
-                    "inputs": ["to_dataframe_step"]
-                },
-                # {
-                #     "name": "column_parser_step",
-                #     "primitives": ["d3m.primitives.data.ColumnParser"],
-                #     "inputs": ["to_dataframe_step"]
-                # },
-                # read X value
-                {
-                    "name": "dataframe_to_tensor",
-                    "primitives": ["d3m.primitives.dsbox.DataFrameToTensor"],
-                    "inputs": ["to_dataframe_step"]
-                },
-                {
-                    "name": "feature_extraction",
-                    "primitives": ["d3m.primitives.dsbox.ResNet50ImageFeature"],
-                    #"primitives": ["d3m.primitives.dsbox.Vgg16ImageFeature"],
-                    "inputs": ["dataframe_to_tensor"]
-                },
-                {
-                    "name": "PCA_step",
-                    "primitives": ["d3m.primitives.sklearn_wrap.SKPCA"],
-                    "inputs": ["feature_extraction"]
-                },
-
-                {
-                    "name": "regressor_step",
-                    "primitives": ["d3m.primitives.sklearn_wrap.SKRandomForestRegressor"],
-                    "inputs": ["PCA_step", "extract_target_step"]
-                },
-            ]
-        }
-
-    def importance(datset, problem_description):
-        return 7
-
-
-class DefaultImageProcessingRegressionTemplate(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "Default_image_processing_regression_template",
-            "taskType": TaskType.REGRESSION.name,
-            # See TaskType, range include 'CLASSIFICATION', 'CLUSTERING', 'COLLABORATIVE_FILTERING', 'COMMUNITY_DETECTION', 'GRAPH_CLUSTERING', 'GRAPH_MATCHING', 'LINK_PREDICTION', 'REGRESSION', 'TIME_SERIES_FORECASTING', 'VERTEX_NOMINATION'
-            "taskSubtype": {TaskSubtype.UNIVARIATE.name, TaskSubtype.MULTIVARIATE.name},
-            "inputType": "image",  # See SEMANTIC_TYPES.keys() for range of values
-            "output": "regressor_step",  # Name of the final step generating the prediction
-            "target": "extract_target_step",  # Name of the step generating the ground truth
-            "steps": [
-
-                {
-                    "name": "denormalize_step",
-                    "primitives": ["d3m.primitives.dsbox.Denormalize"],
-                    "inputs": ["template_input"]
-                },
-                {
-                    "name": "to_dataframe_step",
-                    "primitives": ["d3m.primitives.datasets.DatasetToDataFrame"],
-                    "inputs": ["denormalize_step"]
-                },
-                # read Y value
-                {
-                    "name": "extract_target_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
-                        "hyperparameters":
-                            {
-                                'semantic_types': ('https://metadata.datadrivendiscovery.org/types/TrueTarget',),
-                                'use_columns': (),
-                                'exclude_columns': ()
-                            }
-                        }],
-                    "inputs": ["to_dataframe_step"]
-                },
-                # {
-                #     "name": "column_parser_step",
-                #     "primitives": ["d3m.primitives.data.ColumnParser"],
-                #     "inputs": ["to_dataframe_step"]
-                # },
-                # read X value
-                {
-                    "name": "dataframe_to_tensor",
-                    "primitives": ["d3m.primitives.dsbox.DataFrameToTensor"],
-                    "inputs": ["to_dataframe_step"]
-                },
-                {
-                    "name": "feature_extraction",
-                    "primitives": ["d3m.primitives.dsbox.ResNet50ImageFeature"],
-                    "inputs": ["dataframe_to_tensor"]
-                },
-                {
-                    "name": "PCA_step",
-                    "primitives": ["d3m.primitives.sklearn_wrap.SKPCA"],
-                    "inputs": ["feature_extraction"]
-                },
-
-                {
-                    "name": "regressor_step",
-                    "primitives": ["d3m.primitives.sklearn_wrap.SKRandomForestRegressor"],
-                    "inputs": ["PCA_step", "extract_target_step"]
-                },
-            ]
-        }
-
-    def importance(datset, problem_description):
-        return 7
-
-
-
-################################################################################################################
-#####################################   TextProblemsTemplates   ################################################
-################################################################################################################
-
-
-class DefaultTextClassificationTemplate(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "default_text_classification_template",
-            "taskSubtype": {TaskSubtype.BINARY.name, TaskSubtype.MULTICLASS.name},
-            "taskType": TaskType.CLASSIFICATION.name,
-            # See TaskType, range include 'CLASSIFICATION', 'CLUSTERING', 'COLLABORATIVE_FILTERING',
-            # 'COMMUNITY_DETECTION', 'GRAPH_CLUSTERING', 'GRAPH_MATCHING', 'LINK_PREDICTION',
-            # 'REGRESSION', 'TIME_SERIES_FORECASTING', 'VERTEX_NOMINATION'
-            "inputType": {"text", "table"},  # See SEMANTIC_TYPES.keys() for range of values
-            "output": "model_step",  # Name of the final step generating the prediction
-            "target": "extract_target_step",  # Name of the step generating the ground truth
-            "steps": TemplateSteps.dsbox_generic_text_steps() + [
-                {
-                    "name": "model_step",
-                    "runtime": {
-                        "cross_validation": 10,
-                        "stratified": True
-                    },
-                    "primitives": [
-                        {
-                            "primitive":
-                                "d3m.primitives.sklearn_wrap.SKGradientBoostingClassifier",
-                            "hyperparameters":
-                                {
-                                    'max_depth': [2, 5],
-                                    'n_estimators': [50, 100],
-                                    'learning_rate': [0.1, 0.3],
-                                    'min_samples_split': [2, 3],
-                                    'min_samples_leaf': [1, 2],
-                                }
-                        },
-                        {
-                            "primitive":
-                                "d3m.primitives.sklearn_wrap.SKMultinomialNB",
-                            "hyperparameters":
-                                {
-                                    'alpha': [0, .5, 1],
-                                }
-                        },
-                        {
-                            "primitive":
-                                "d3m.primitives.sklearn_wrap.SKRandomForestClassifier",
-                            "hyperparameters":
-                                {
-                                }
-                        },
-                    ],
-                    "inputs": ["data", "target"]
-                }
-            ]
-        }
-
-    # @override
-    def importance(datset, problem_description):
-        return 7
-
-
-class DefaultTextRegressionTemplate(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "default_text_regression_template",
-            "taskSubtype": {TaskSubtype.UNIVARIATE.name, TaskSubtype.MULTIVARIATE.name},
-            "taskType": TaskType.REGRESSION.name,
-            # See TaskType, range include 'CLASSIFICATION', 'CLUSTERING', 'COLLABORATIVE_FILTERING',
-            # 'COMMUNITY_DETECTION', 'GRAPH_CLUSTERING', 'GRAPH_MATCHING', 'LINK_PREDICTION',
-            # 'REGRESSION', 'TIME_SERIES_FORECASTING', 'VERTEX_NOMINATION'
-            "inputType": {"text", "table"},  # See SEMANTIC_TYPES.keys() for range of values
-            "output": "model_step",  # Name of the final step generating the prediction
-            "target": "extract_target_step",  # Name of the step generating the ground truth
-            "steps": TemplateSteps.dsbox_generic_text_steps() + [
-                {
-                    "name": "model_step",
-                    "runtime": {
-                        "cross_validation": 10,
-                    },
-                    "primitives": [
-                        {
-                            "primitive":
-                                "d3m.primitives.sklearn_wrap.SKGradientBoostingRegressor",
-                            "hyperparameters":
-                                {
-                                    'max_depth': [2, 5],
-                                    'n_estimators': [100, 200],
-                                    'learning_rate': [0.1, 0.3],
-                                    'min_samples_split': [2, 3],
-                                    'min_samples_leaf': [1, 2],
-                                }
-                        },
-                        {
-                            "primitive":
-                                "d3m.primitives.sklearn_wrap.SKRandomForestRegressor",
-                            "hyperparameters":
-                                {
-                                }
-                        },
-                    ],
-                    "inputs": ["data", "target"]
-                }
-            ]
-        }
-
-    # @override
-    def importance(datset, problem_description):
-        return 7
-
-
-################################################################################################################
-#####################################   GraphProblemsTemplates   ###############################################
-################################################################################################################
-
-
-class SRILinkPredictionTemplate(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "SRI_LinkPrediction_Template",
-            "taskType": {TaskType.LINK_PREDICTION.name,TaskType.GRAPH_MATCHING.name, TaskType.VERTEX_NOMINATION.name},
-            "taskSubtype": "NONE",
-            "inputType": "graph",
-            "output": "model_step",
-            "steps": [
-                {
-                    "name": "denormalize_step",
-                    "primitives": ["d3m.primitives.datasets.Denormalize"],
-                    "inputs": ["template_input"]
-                },
-                {
-                    "name":"to_dataframe_step",
-                    "primitives":["d3m.primitives.datasets.DatasetToDataFrame"],
-                    "inputs":["denormalize_step"]
-                },
-                {
-                    "name": "extract_attribute_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
-                        "hyperparameters":
-                            {
-                                'semantic_types': ('https://metadata.datadrivendiscovery.org/types/Attribute',),
-                                'use_columns': (),
-                                'exclude_columns': ()
-                            }
-                        }],
-                    "inputs": ["to_dataframe_step"]
-                },
-                {
-                    "name": "extract_target_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
-                        "hyperparameters":
-                            {
-                                'semantic_types': ('https://metadata.datadrivendiscovery.org/types/TrueTarget',),
-                                'use_columns': (),
-                                'exclude_columns': ()
-                            }
-                        }],
-                    "inputs": ["to_dataframe_step"]
-                },
-                {
-                    "name": "model_step",
-                    "primitives":[{
-                        "primitive":"d3m.primitives.sklearn_wrap.SKRandomForestClassifier",
-                        "hyperparameters":{
-                                    # 'bootstrap': [True, False],
-                                    'max_depth': [15, 30, None],
-                                    'min_samples_leaf': [1, 2, 4],
-                                    'min_samples_split': [2, 5, 10],
-                                    'max_features': ['auto', 'sqrt'],
-                                    'n_estimators': [10, 50, 100],
-                            }
-                        }
-                    ],
-                    "inputs":["extract_attribute_step", "extract_target_step"]
-                }
-            ]
-        }
-
-    def importance(datset, problem_description):
-        return 7
-
-
-class SRIGraphMatchingTemplate(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "SRI_GraphMatching_Template",
-            "taskType": {TaskType.GRAPH_MATCHING.name, TaskType.LINK_PREDICTION.name},
-            # for some special condition, the taskSubtype can be "NONE" which indicate no taskSubtype given
-            "taskSubtype": "NONE",
-            "inputType": "graph",
-            "output": "model_step",
-            "steps": [
-                {
-                    "name": "model_step",
-                    "primitives": [
-                        {
-                            "primitive":"d3m.primitives.sri.psl.GraphMatchingLinkPrediction",
-                            "hyperparameters":{
-                                "link_prediction_hyperparams" : [(TemplateSteps.class_hyperparameter_generator("d3m.primitives.sri.psl.GraphMatchingLinkPrediction", "link_prediction_hyperparams", {"truth_threshold": 0.0000001,"psl_options": "","psl_temp_dir": "/tmp/psl/run","postgres_db_name": "psl_d3m","admm_iterations": 1000,"max_threads": 0,"jvm_memory": 0.75,"prediction_column": "match"}))]
-                                }
-                        }
-                    ],
-                    "inputs": ["template_input"]
-                }
-            ]
-        }
-
-    def importance(dataset, problem_description):
-        return 7
-
-
-class SRIVertexNominationTemplate(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "SRI_Vertex_Nomination_Template",
-            "taskType": TaskType.VERTEX_NOMINATION.name,
-            "taskSubtype": "NONE",
-            "inputType": {"graph", "edgeList"},
-            "output": "model_step",
-            "steps": [
-                {
-                    "name": "parse_step",
-                    "primitives": ["d3m.primitives.sri.graph.VertexNominationParser"],
-                    "inputs": ["template_input"]
-
-                },
-                {
-                    "name": "model_step",
-                    "primitives": ["d3m.primitives.sri.psl.VertexNomination"],
-                    "inputs": ["parse_step"]
-
-                }
-            ]
-        }
-
-    def importance(datset, problem_description):
-        return 7
-
-
-class SRICollaborativeFilteringTemplate(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "SRI_Collaborative_Filtering_Template",
-            "taskType": {TaskType.COLLABORATIVE_FILTERING.name},
-            "taskSubtype": "NONE",
-            "inputType": "table",
-            "output": "model_step",
-            "steps": [
-                {
-                    "name": "model_step",
-                    "primitives": ["d3m.primitives.sri.psl.CollaborativeFilteringLinkPrediction"],
-                    "inputs": ["template_input"]
-
-                }
-            ]
-        }
-
-    def importance(dataset, problem_description):
-        return 7
-
-
-class SRICommunityDetectionTemplate(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "SRI_Community_Detection_Template",
-            "taskType": {TaskType.COMMUNITY_DETECTION.name},
-            "taskSubtype": {TaskSubtype.NONOVERLAPPING.name, TaskSubtype.OVERLAPPING.name},
-            "inputType": "graph",
-            "output": "model_step",
-            "steps": [
-                {
-                    "name": "parser_step",
-                    "primitives": ["d3m.primitives.sri.graph.CommunityDetectionParser"],
-                    "inputs": ["template_input"]
-
-                },
-                {
-                    "name": "model_step",
-                    "primitives": ["d3m.primitives.sri.psl.CommunityDetection"],
-                    "inputs": ["parser_step"]
-                }
-            ]
-        }
-
-    def importance(dataset, problem_description):
-        return 7
-
-
-'''
-JHU needs R supports and their primitives seem to be failing in pickling
-'''
-class JHUVertexNominationTemplate(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "JHU_Vertex_Nomination_Template",
-            "taskType": TaskType.VERTEX_NOMINATION.name,
-            "taskSubtype": "NONE",
-            "inputType": "graph",
-            "output": "model_step",
-            "steps": [
-                {
-                    "name": "model_step",
-                    "primitives": ["d3m.primitives.jhu_primitives.SpectralGraphClustering"],
-                    "inputs": ["template_input"]
-
-                }
-            ]
-        }
-
-    def importance(datset, problem_description):
-        return 7
-
-
-class JHUGraphMatchingTemplate(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "JHU_Graph_Matching_Template",
-            "taskType": TaskType.GRAPH_MATCHING.name,
-            "taskSubtype": "NONE",
-            "inputType": "graph",
-            "output": "model_step",
-            "steps": [
-                {
-                    "name": "model_step",
-                    "primitives": ["d3m.primitives.jhu_primitives.SeededGraphMatching"],
-                    "inputs": ["template_input"]
-                }
-            ]
-        }
-
-    def importance(datset, problem_description):
-        return 7
-
-
-################################################################################################################
-#####################################   AudioClassificationTemplate   ##########################################
-################################################################################################################
-
-
-class BBNAudioClassificationTemplate(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "BBN_Audio_Classification_Template",
-            "taskType": {TaskType.CLASSIFICATION.name},
-            "taskSubtype": {TaskSubtype.MULTICLASS.name},
-            "inputType": "audio",
-            "output": "model_step",
-            "steps": [
-                {
-                    "name": "denormalize_step",
-                    "primitives": ["d3m.primitives.dsbox.Denormalize"],
-                    "inputs": ["template_input"]
-                },
-                {
-                    "name": "to_dataframe_step",
-                    "primitives": ["d3m.primitives.datasets.DatasetToDataFrame"],
-                    "inputs": ["denormalize_step"]
-                },
-                {
-                    "name": "readtarget_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
-                        "hyperparameters":
-                            {
-                                'semantic_types': ('https://metadata.datadrivendiscovery.org/types/TrueTarget',),
-                                'use_columns': (),
-                                'exclude_columns': ()
-                            }
-                        }],
-                    "inputs": ["to_dataframe_step"]
-                },
-                {
-                    "name": "readaudio_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.bbn.time_series.AudioReader",
-                        "hyperparameters":
-                            {
-                                "read_as_mono": [(True)],
-                                "resampling_rate": [(16000.0)],
-                            }
-                    }],
-                    "inputs": ["template_input"]
-                },
-                {
-                    "name": "channel_step",
-                    "primitives": ["d3m.primitives.bbn.time_series.ChannelAverager"],
-                    "inputs": ["readaudio_step"]
-                },
-                {
-                    "name": "signaldither_step",
-                    "primitives": [{"primitive": "d3m.primitives.bbn.time_series.SignalDither",
-                                    "hyperparameters": {
-                                        "level": [(0.0001)],
-                                        "reseed": [(True)]
-                                    }
-                                    }],
-                    "inputs": ["channel_step"]
-                },
-                {
-                    "name": "signalframer_step",
-                    "primitives": [{"primitive": "d3m.primitives.bbn.time_series.SignalFramer",
-                                    "hyperparameters": {
-                                        "flatten_output": [(False)],
-                                        "frame_length_s": [(0.025)],
-                                        "frame_shift_s": [(0.01)]
-                                    }
-                                    }],
-                    "inputs": ["signaldither_step"]
-                },
-                {
-                    "name": "MFCC_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.bbn.time_series.SignalMFCC",
-                        "hyperparameters": {
-                            "cep_lifter": [(22.0)],
-                            "frame_mean_norm": [(False)],
-                            "nfft": [(None)],
-                            "num_ceps": [(20)],
-                            "num_chans": [(20)],
-                            "preemcoef": [(None)],
-                            "use_power": [(False)]
-                        }
-                    }],
-                    "inputs": ["signalframer_step"]
-                },
-                {
-                    "name": "vectorextractor_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.bbn.time_series.IVectorExtractor",
-                        "hyperparameters": {
-                            "gmm_covariance_type": [("diag")],
-                            "ivec_dim": [(100)],
-                            "max_gmm_iter": [(20)],
-                            "num_gauss": [(32)],
-                            "num_ivec_iter": [(7)]
-                        }
-                    }],
-                    "inputs": ["MFCC_step"]
-                },
-                {
-                    "name": "model_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.bbn.sklearn_wrap.BBNMLPClassifier",
-                        "hyperparameters": {
-                            "activation": [("relu")],
-                            "add_index_columns": [(True)],
-                            "alpha": [(0.0001)],
-                            "beta_1": [(0.9)],
-                            "beta_2": [(0.999)],
-                            "early_stopping": [(True)],
-                            "epsilon": [(1e-8)],
-                            "exclude_columns": [([])],
-                            # "hidden_layer_sizes":[([30,30])],
-                            "learning_rate": [("constant")],
-                            "learning_rate_init": [(0.01)],
-                            "max_iter": [(200)],
-                            "return_result": [("replace")],
-                            "shuffle": [(True)],
-                            "solver": [("adam")],
-                            "tol": [(0.0001)],
-                            "use_columns": [([])],
-                            "use_semantic_types": [(False)],
-                            "warm_start": [(False)]
-                        }
-                    }],
-                    "inputs": ["vectorextractor_step", "readtarget_step"]
-
-                }
-            ]
-        }
-
-    def importance(datset, problem_description):
-        return 7
-
-
-################################################################################################################
-#####################################   SRIMeanbaselineTemplate   ##############################################
-################################################################################################################
-
-
-class SRIMeanBaselineTemplate(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "SRI_Mean_Baseline_Template",
-            "taskType": "NONE",
-            "taskSubtype": "NONE",
-            "inputType": "NONE",
-            "output": "model_step",
-            "steps": [
-                {
-                    "name": "model_step",
-                    "primitives": ["d3m.primitives.sri.baseline.MeanBaseline"],
-                    "inputs": ["template_input"]
-
-                }
-            ]
-        }
-
-    def importance(dataset, problem_description):
-        return 10
-
-
-################################################################################################################
-#####################################   ClusteringTemplate   ###################################################
-################################################################################################################
-
-
-class CMUClusteringTemplate(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "CMU_Clustering_Template",
-            "taskType": TaskType.CLUSTERING.name,
-            "taskSubtype": "NONE",
-            "inputType": "table",
-            "output": "model_step",
-            "steps": [
-                {
-                    "name": "denormalize_step",
-                    "primitives": ["d3m.primitives.dsbox.Denormalize"],
-                    "inputs": ["template_input"]
-                },
-                {
-                    "name": "to_dataframe_step",
-                    "primitives": ["d3m.primitives.datasets.DatasetToDataFrame"],
-                    "inputs": ["denormalize_step"]
-                },
-                {
-                    "name": "column_parser_step",
-                    "primitives": ["d3m.primitives.data.ColumnParser"],
-                    "inputs": ["to_dataframe_step"]
-                },
-                {
-                    "name": "extract_attribute_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
-                        "hyperparameters":
-                            {
-                                'semantic_types': (
-                                    'https://metadata.datadrivendiscovery.org/types/Attribute',),
-                                'use_columns': (),
-                                'exclude_columns': ()
-                            }
-                    }],
-                    "inputs": ["column_parser_step"]
-                },
-                {
-                    "name": "model_step",
-                    "primitives": [
-                        {
-                            "primitive": "d3m.primitives.cmu.fastlvm.GMM",
-                            "hyperparameters": {
-                                "k": [(4), (6), (8), (10), (12)]
-                            }
-                        }
-                    ],
-                    "inputs": ["extract_attribute_step"]
-                }
-            ]
-        }
-
-    def importance(datset, problem_description):
-        return 7
-
-
-################################################################################################################
-#####################################   VideoClassificationTemplate   ##########################################
-################################################################################################################
-
-
-class MichiganVideoClassificationTemplate(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "Michigan_Video_Classification_Template",
-            "taskType": TaskType.CLASSIFICATION.name,
-            # See TaskType, range include 'CLASSIFICATION', 'CLUSTERING', 'COLLABORATIVE_FILTERING',
-            # 'COMMUNITY_DETECTION', 'GRAPH_CLUSTERING', 'GRAPH_MATCHING', 'LINK_PREDICTION',
-            # 'REGRESSION', 'TIME_SERIES_FORECASTING', 'VERTEX_NOMINATION'
-            "taskSubtype": TaskSubtype.MULTICLASS.name,
-            "inputType": "video",  # See SEMANTIC_TYPES.keys() for range of values
-            "output": "model_step",  # Name of the final step generating the prediction
-            "target": "extract_target_step",  # Name of the step generating the ground truth
-            "steps": [
-                {
-                    "name": "denormalize_step",
-                    "primitives": ["d3m.primitives.dsbox.Denormalize"],
-                    "inputs": ["template_input"]
-                },
-                {
-                    "name": "to_dataframe_step",
-                    "primitives": ["d3m.primitives.datasets.DatasetToDataFrame"],
-                    "inputs": ["denormalize_step"]
-                },
-                {
-                    "name": "extract_target_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
-                        "hyperparameters":
-                            {
-                                'semantic_types': (
-                                    'https://metadata.datadrivendiscovery.org/types/TrueTarget',),
-                                'use_columns': (),
-                                'exclude_columns': ()
-                            }
-                    }],
-                    "inputs": ["to_dataframe_step"]
-                },
-                {
-                    "name": "read_video_step",
-                    "primitives": ["d3m.primitives.data.VideoReader"],
-                    "inputs": ["to_dataframe_step"]
-                },
-                {
-                    "name": "featurize_step",
-                    "primitives": ["d3m.primitives.spider.featurization.I3D"],
-                    "inputs": ["read_video_step"]
-
-                },
-                {
-                    "name": "convert_step",
-                    "primitives": ["d3m.primitives.data.NDArrayToDataFrame"],
-                    "inputs": ["featurize_step"]
-
-                },
-                {
-                    "name": "model_step",
-                    "primitives": ["d3m.primitives.classifier.RandomForest"],
-                    "inputs": ["convert_step", "extract_target_step"]
-                },
-            ]
-        }
-
-    # @override
-    def importance(datset, problem_description):
-        return 7
-
-
-################################################################################################################
-#####################################   TA1Template   ##########################################################
-################################################################################################################
-
-class TA1ClassificationTemplate1(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "TA1_classification_template_1",
-            "taskSubtype": {TaskSubtype.BINARY.name, TaskSubtype.MULTICLASS.name},
-            "taskType": TaskType.CLASSIFICATION.name,
-            "inputType": "table",  # See SEMANTIC_TYPES.keys() for range of values
-            "output": "model_step",  # Name of the final step generating the prediction
-            "target": "extract_target_step",  # Name of the step generating the ground truth
-            "steps": [
-                {
-                    "name": "denormalize_step",
-                    "primitives": ["d3m.primitives.dsbox.Denormalize"],
-                    "inputs": ["template_input"]
-                },
-                {
-                    "name": "to_dataframe_step",
-                    "primitives": ["d3m.primitives.datasets.DatasetToDataFrame"],
-                    "inputs": ["denormalize_step"]
-                },
-                {
-                    "name": "extract_attribute_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
-                        "hyperparameters":
-                            {
-                                'semantic_types': (
-                                    'https://metadata.datadrivendiscovery.org/types/Attribute',),
-                                'use_columns': (),
-                                'exclude_columns': ()
-                            }
-                    }],
-                    "inputs": ["to_dataframe_step"]
-                },
-                {
-                    "name": "extract_target_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
-                        "hyperparameters":
-                            {
-                                'semantic_types': ('https://metadata.datadrivendiscovery.org/types/TrueTarget',),
-                                'use_columns': (),
-                                'exclude_columns': ()
-                            }
-                        }],
-                    "inputs": ["to_dataframe_step"]
-                },
-                {
-                    "name": "encode_step",
-                    "primitives": ["d3m.primitives.dsbox.Encoder"],
-                    "inputs": ["extract_attribute_step"]
-                },
-                {
-                    "name": "impute_step",
-                    "primitives": ["d3m.primitives.dsbox.MeanImputation"],
-                    "inputs": ["encode_step"]
-                },
-                {
-                    "name": "cast_1_step",
-                    "primitives": [
-                        {
-                            "primitive": "d3m.primitives.data.CastToType",
-                            "hyperparameters": {"type_to_cast": ["float"]}
-                        },
-                        "d3m.primitives.dsbox.DoNothing",
-                    ],
-                    "inputs": ["impute_step"]
-                },
-                {
-                    "name": "model_step",
-                    "runtime": {
-                        "cross_validation": 10,
-                        "stratified": True
-                    },
-                    "primitives": [{
-                        "primitive":
-                            "d3m.primitives.sklearn_wrap.SKRandomForestClassifier",
-                        "hyperparameters":
-                            {
-                                'max_depth': [(2), (4), (8)],  # (10), #
-                                'n_estimators': [(10), (20), (30)]
-                            }
-                    },
-                    ],
-                    "inputs": ["cast_1_step", "extract_target_step"]
-                }
-            ]
-        }
-
-    # @override
-    def importance(datset, problem_description):
-        return 7
-
-
-class TA1Classification_2(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "TA1Classification_2",
-            "taskSubtype": {TaskSubtype.BINARY.name, TaskSubtype.MULTICLASS.name},
-            "taskType": TaskType.CLASSIFICATION.name,
-            # See TaskType, range include 'CLASSIFICATION', 'CLUSTERING', 'COLLABORATIVE_FILTERING',
-            # 'COMMUNITY_DETECTION', 'GRAPH_CLUSTERING', 'GRAPH_MATCHING', 'LINK_PREDICTION',
-            # 'REGRESSION', 'TIME_SERIES_FORECASTING', 'VERTEX_NOMINATION'
-            "inputType": "text",  # See SEMANTIC_TYPES.keys() for range of values
-            "output": "model_step",  # Name of the final step generating the prediction
-            "target": "extract_target_step",  # Name of the step generating the ground truth
-            "steps": [
-                *default_dataparser(attribute_name="extract_attribute_step",
-                                    target_name="extract_target_step"),
-                {
-                    "name": "corex_step",
-                    "primitives": ["d3m.primitives.dsbox.CorexText"],
-                    "inputs": ["extract_attribute_step"]
-                },
-                {
-                    "name": "encoder_step",
-                    "primitives": ["d3m.primitives.dsbox.Labler"],
-                    "inputs": ["corex_step"]
-                },
-                {
-                    "name": "impute_step",
-                    "primitives": ["d3m.primitives.dsbox.MeanImputation"],
-                    "inputs": ["encoder_step"]
-                },
-                {
-                    "name": "nothing_step",
-                    "primitives": ["d3m.primitives.dsbox.DoNothing"],
-                    "inputs": ["impute_step"]
-                },
-                {
-                    "name": "scaler_step",
-                    "primitives": ["d3m.primitives.dsbox.IQRScaler"],
-                    "inputs": ["nothing_step"]
-                },
-                {
-                    "name": "model_step",
-                    "runtime": {
-                        "cross_validation": 10,
-                        "stratified": True
-                    },
-                    "primitives": ["d3m.primitives.sklearn_wrap.SKRandomForestClassifier"],
-                    "inputs": ["scaler_step", "extract_target_step"]
-                }
-            ]
-        }
-        # import pprint
-        # pprint.pprint(self.template)
-        # exit(1)
-
-    # @override
-    def importance(datset, problem_description):
-        return 7
-
-
-class TA1Classification_3(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "TA1Classification_3",
-            "taskSubtype": {TaskSubtype.BINARY.name, TaskSubtype.MULTICLASS.name},
-            "taskType": TaskType.CLASSIFICATION.name,
-            "inputType": "table",  # See SEMANTIC_TYPES.keys() for range of values
-            "output": "model_step",  # Name of the final step generating the prediction
-            "target": "extract_target_step",  # Name of the step generating the ground truth
-            "steps": [
-                {
-                    "name": "denormalize_step",
-                    "primitives": ["d3m.primitives.dsbox.Denormalize"],
-                    "inputs": ["template_input"]
-                },
-                {
-                    "name": "to_dataframe_step",
-                    "primitives": ["d3m.primitives.datasets.DatasetToDataFrame"],
-                    "inputs": ["denormalize_step"]
-                },
-                {
-                    "name": "extract_attribute_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
-                        "hyperparameters":
-                            {
-                                'semantic_types': ('https://metadata.datadrivendiscovery.org/types/Attribute',),
-                                'use_columns': (),
-                                'exclude_columns': ()
-                            }
-                    }],
-                    "inputs": ["to_dataframe_step"]
-                },
-                {
-                    "name": "extract_target_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
-                        "hyperparameters":
-                            {
-                                'semantic_types': ('https://metadata.datadrivendiscovery.org/types/TrueTarget',),
-                                'use_columns': (),
-                                'exclude_columns': ()
-                            }
-                        }],
-                    "inputs": ["to_dataframe_step"]
-                },
-                {
-                    "name": "profile_step",
-                    "primitives": ["d3m.primitives.dsbox.Profiler"],
-                    "inputs": ["extract_attribute_step"]
-                },
-                {
-                    "name": "clean_step",
-                    "primitives": [
-                        "d3m.primitives.dsbox.CleaningFeaturizer",
-                        "d3m.primitives.dsbox.DoNothing",
-                    ],
-                    "inputs": ["profiler_step"]
-                },
-                {
-                    "name": "impute_step",
-                    "primitives": ["d3m.primitives.dsbox.MeanImputation"],
-                    "inputs": ["clean_step"]
-                },
-                {
-                    "name": "label_step",
-                    "primitives": ["d3m.primitives.dsbox.Labler"],
-                    "inputs": ["impute_step"]
-                },
-                {
-                    "name": "corex_step",
-                    "primitives": ["d3m.primitives.dsbox.CorexText"],
-                    "inputs": ["label_step"]
-                },
-                # {
-                #     "name": "corex_step",
-                #     "primitives": ["d3m.primitives.dsbox.CorexText"],
-                #     "inputs": ["cast_1_step"]
-                # },
-                {
-                    "name": "model_step",
-                    "primitives": [
-                        "d3m.primitives.sklearn_wrap.SKRandomForestClassifier"],
-                    "runtime": {
-                        "cross_validation": 10,
-                        "stratified": False
-                    },
-                    "inputs": ["corex_step", "extract_target_step"]
-                }
-            ]
-        }
-
-    # @override
-
-    def importance(datset, problem_description):
-        return 7
-
-
-class MuxinTA1ClassificationTemplate1(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "MuxinTA1ClassificationTemplate1",
-            "taskSubtype": {TaskSubtype.BINARY.name, TaskSubtype.MULTICLASS.name},
-            "taskType": TaskType.CLASSIFICATION.name,
-            "inputType": "table",  # See SEMANTIC_TYPES.keys() for range of values
-            "output": "model_step",  # Name of the final step generating the prediction
-            "target": "extract_target_step",  # Name of the step generating the ground truth
-            "steps": [
-                {
-                    "name": "denormalize_step",
-                    "primitives": ["d3m.primitives.dsbox.Denormalize"],
-                    "inputs": ["template_input"]
-                },
-                {
-                    "name": "to_dataframe_step",
-                    "primitives": ["d3m.primitives.datasets.DatasetToDataFrame"],
-                    "inputs": ["denormalize_step"]
-                },
-                {
-                    "name": "extract_attribute_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
-                        "hyperparameters":
-                            {
-                                'semantic_types': ('https://metadata.datadrivendiscovery.org/types/Attribute',),
-                                'use_columns': (),
-                                'exclude_columns': ()
-                            }
-                    }],
-                    "inputs": ["to_dataframe_step"]
-                },
-                {
-                    "name": "extract_target_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
-                        "hyperparameters":
-                            {
-                                'semantic_types': ('https://metadata.datadrivendiscovery.org/types/TrueTarget',),
-                                'use_columns': (),
-                                'exclude_columns': ()
-                            }
-                        }],
-                    "inputs": ["to_dataframe_step"]
-                },
-                {
-                    "name": "encode1_step",
-                    # "primitives": ["d3m.primitives.dsbox.UnaryEncoder","d3m.primitives.dsbox.DoNothing"],
-                    "primitives": ["d3m.primitives.dsbox.UnaryEncoder"],
-                    "inputs": ["extract_attribute_step"]
-                },
-                {
-                    "name": "encode2_step",
-                    "primitives": ["d3m.primitives.dsbox.Encoder", "d3m.primitives.dsbox.DoNothing"],
-                    # "primitives": ["d3m.primitives.dsbox.Encoder"],
-                    "inputs": ["encode1_step"]
-                },
-                {
-                    "name": "impute_step",
-                    "primitives": ["d3m.primitives.dsbox.MeanImputation"],
-                    "inputs": ["encode2_step"]
-                },
-                {
-                    "name": "cast_1_step",
-                    "primitives": [
-                        {
-                            "primitive": "d3m.primitives.data.CastToType",
-                            "hyperparameters": {"type_to_cast": ["float"]}
-                        },
-                        "d3m.primitives.dsbox.DoNothing",
-                    ],
-                    "inputs": ["impute_step"]
-                },
-                {
-                    "name": "model_step",
-                    "runtime": {
-                        "cross_validation": 2,
-                        # "test_validation":1,
-                        "stratified": True
-                    },
-                    "primitives": [
-                        {
-                            "primitive":
-                                "d3m.primitives.sklearn_wrap.SKRandomForestClassifier",
-                            "hyperparameters":
-                                {
-                                    'max_depth': [(2), (4)],  # (10), #
-                                    'n_estimators': [(10), (30)]
-                                }
-                        },
-                        # {
-                        #     "primitive":
-                        #         "d3m.primitives.sklearn_wrap.SKLinearSVC",
-                        #     "hyperparameters":
-                        #         {
-                        #             'C': [(1), (10)],  # (10), #
-                        #         }
-                        # },
-                        # {
-                        #     "primitive":
-                        #         "d3m.primitives.sklearn_wrap.SKMultinomialNB",
-                        #     "hyperparameters":
-                        #         {
-                        #             'alpha': [(1)],
-                        #         }
-                        # },
-                    ],
-                    "inputs": ["cast_1_step", "extract_target_step"]
-                }
-            ]
-        }
-
-    # @override
-
-    def importance(datset, problem_description):
-        return 7
-
-
-class MuxinTA1ClassificationTemplate2(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "MuxinTA1ClassificationTemplate2",
-            "taskSubtype": {TaskSubtype.BINARY.name, TaskSubtype.MULTICLASS.name},
-            "taskType": TaskType.CLASSIFICATION.name,
-            "inputType": "table",  # See SEMANTIC_TYPES.keys() for range of values
-            "output": "model_step",  # Name of the final step generating the prediction
-            "target": "extract_target_step",  # Name of the step generating the ground truth
-            "steps": [
-                {
-                    "name": "denormalize_step",
-                    "primitives": ["d3m.primitives.dsbox.Denormalize"],
-                    "inputs": ["template_input"]
-                },
-                {
-                    "name": "to_dataframe_step",
-                    "primitives": ["d3m.primitives.datasets.DatasetToDataFrame"],
-                    "inputs": ["denormalize_step"]
-                },
-                {
-                    "name": "extract_attribute_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
-                        "hyperparameters":
-                            {
-                                'semantic_types': ('https://metadata.datadrivendiscovery.org/types/Attribute',),
-                                'use_columns': (),
-                                'exclude_columns': ()
-                            }
-                    }],
-                    "inputs": ["to_dataframe_step"]
-                },
-                {
-                    "name": "no_op_step",
-                    "primitives": ["d3m.primitives.dsbox.DoNothing"],
-                    "inputs": ["to_dataframe_step"]
-                },
-                {
-                    "name": "extract_target_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
-                        "hyperparameters":
-                            {
-                                'semantic_types': ('https://metadata.datadrivendiscovery.org/types/TrueTarget',),
-                                'use_columns': (),
-                                'exclude_columns': ()
-                            }
-                        }],
-                    "inputs": ["to_dataframe_step"]
-                },
-                {
-                    "name": "encode_step",
-                    "primitives": ["d3m.primitives.dsbox.Encoder"],
-                    "inputs": ["extract_attribute_step"]
-                },
-                {
-                    "name": "impute_step",
-                    "primitives": ["d3m.primitives.dsbox.GreedyImputation"],
-                    "inputs": ["encode_step", "extract_target_step"]
-                },
-                # {
-                #     "name": "corex_step",
-                #     "primitives": ["d3m.primitives.dsbox.CorexText"],
-                #     "inputs": ["cast_1_step"]
-                # },
-                {
-                    "name": "cast_1_step",
-                    "primitives": [
-                        {
-                            "primitive": "d3m.primitives.data.CastToType",
-                            "hyperparameters": {"type_to_cast": ["float"]}
-                        },
-                        {
-                            "primitive": "d3m.primitives.dsbox.DoNothing",
-                            "hyperparameters": {}
-                        }
-                    ],
-                    "inputs": ["impute_step"]
-                },
-                {
-                    "name": "model_step",
-                    "primitives": ["d3m.primitives.sklearn_wrap.SKRandomForestClassifier"],
-                    "runtime": {
-                        "cross_validation": 10,
-                        "stratified": False
-                    },
-                    "inputs": ["cast_1_step", "extract_target_step"]
-                }
-            ]
-        }
-
-    # @override
-
-    def importance(datset, problem_description):
-        return 7
-
-
-class MuxinTA1ClassificationTemplate3(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "MuxinTA1ClassificationTemplate3",
-            "taskSubtype": {TaskSubtype.BINARY.name, TaskSubtype.MULTICLASS.name},
-            "taskType": TaskType.CLASSIFICATION.name,
-            "inputType": "table",  # See SEMANTIC_TYPES.keys() for range of values
-            "output": "model_step",  # Name of the final step generating the prediction
-            "target": "extract_target_step",  # Name of the step generating the ground truth
-            "steps": [
-                {
-                    "name": "denormalize_step",
-                    "primitives": ["d3m.primitives.dsbox.Denormalize"],
-                    "inputs": ["template_input"]
-                },
-                {
-                    "name": "to_dataframe_step",
-                    "primitives": ["d3m.primitives.datasets.DatasetToDataFrame"],
-                    "inputs": ["denormalize_step"]
-                },
-                {
-                    "name": "extract_attribute_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
-                        "hyperparameters":
-                            {
-                                'semantic_types': ('https://metadata.datadrivendiscovery.org/types/Attribute',),
-                                'use_columns': (),
-                                'exclude_columns': ()
-                            }
-                    }],
-                    "inputs": ["to_dataframe_step"]
-                },
-                {
-                    "name": "no_op_step",
-                    "primitives": ["d3m.primitives.dsbox.DoNothing"],
-                    "inputs": ["to_dataframe_step"]
-                },
-                {
-                    "name": "extract_target_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
-                        "hyperparameters":
-                            {
-                                'semantic_types': ('https://metadata.datadrivendiscovery.org/types/TrueTarget',),
-                                'use_columns': (),
-                                'exclude_columns': ()
-                            }
-                        }],
-                    "inputs": ["no_op_step"]
-                },
-                {
-                    "name": "encode_step",
-                    "primitives": ["d3m.primitives.dsbox.Encoder"],
-                    "inputs": ["extract_attribute_step"]
-                },
-                {
-                    "name": "impute_step",
-                    "primitives": ["d3m.primitives.dsbox.IterativeRegressionImputation"],
-                    "inputs": ["encode_step"]
-                },
-                # {
-                #     "name": "corex_step",
-                #     "primitives": ["d3m.primitives.dsbox.CorexText"],
-                #     "inputs": ["cast_1_step"]
-                # },
-                {
-                    "name": "cast_1_step",
-                    "primitives": [
-                        {
-                            "primitive": "d3m.primitives.data.CastToType",
-                            "hyperparameters": {"type_to_cast": ["float"]}
-                        },
-                        {
-                            "primitive": "d3m.primitives.dsbox.DoNothing",
-                            "hyperparameters": {}
-                        }
-                    ],
-                    "inputs": ["impute_step"]
-                },
-                {
-                    "name": "model_step",
-                    "primitives": ["d3m.primitives.sklearn_wrap.SKRandomForestClassifier"],
-                    "runtime": {
-                        "cross_validation": 10,
-                        "stratified": False
-                    },
-                    "inputs": ["cast_1_step", "extract_target_step"]
-                }
-            ]
-        }
-
-    # @override
-
-    def importance(datset, problem_description):
-        return 7
-
-
-class MuxinTA1ClassificationTemplate4(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "MuxinTA1ClassificationTemplate4",
-            "taskSubtype": {TaskSubtype.BINARY.name, TaskSubtype.MULTICLASS.name},
-            "taskType": TaskType.CLASSIFICATION.name,
-            "inputType": "table",  # See SEMANTIC_TYPES.keys() for range of values
-            "output": "model_step",  # Name of the final step generating the prediction
-            "target": "extract_target_step",  # Name of the step generating the ground truth
-            "steps": [
-                {
-                    "name": "denormalize_step",
-                    "primitives": ["d3m.primitives.dsbox.Denormalize"],
-                    "inputs": ["template_input"]
-                },
-                {
-                    "name": "to_dataframe_step",
-                    "primitives": ["d3m.primitives.datasets.DatasetToDataFrame"],
-                    "inputs": ["denormalize_step"]
-                },
-                {
-                    "name": "extract_attribute_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
-                        "hyperparameters":
-                            {
-                                'semantic_types': ('https://metadata.datadrivendiscovery.org/types/Attribute',),
-                                'use_columns': (),
-                                'exclude_columns': ()
-                            }
-                    }],
-                    "inputs": ["to_dataframe_step"]
-                },
-                {
-                    "name": "extract_target_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
-                        "hyperparameters":
-                            {
-                                'semantic_types': ('https://metadata.datadrivendiscovery.org/types/TrueTarget',),
-                                'use_columns': (),
-                                'exclude_columns': ()
-                            }
-                        }],
-                    "inputs": ["to_dataframe_step"]
-                },
-                {
-                    "name": "encode1_step",
-                    "primitives": ["d3m.primitives.dsbox.UnaryEncoder"],
-                    "inputs": ["extract_attribute_step"]
-                },
-                {
-                    "name": "encode2_step",
-                    "primitives": ["d3m.primitives.dsbox.Encoder"],
-                    "inputs": ["encode1_step"]
-                },
-                {
-                    "name": "impute_step",
-                    "primitives": ["d3m.primitives.dsbox.MeanImputation"],
-                    "inputs": ["encode2_step"]
-                },
-                {
-                    "name": "cast_1_step",
-                    "primitives": [
-                        {
-                            "primitive": "d3m.primitives.data.CastToType",
-                            "hyperparameters": {"type_to_cast": ["float"]}
-                        },
-                        "d3m.primitives.dsbox.DoNothing",
-                    ],
-                    "inputs": ["impute_step"]
-                },
-                {
-                    "name": "model_step",
-                    "primitives": [
-                        "d3m.primitives.sklearn_wrap.SKRandomForestClassifier"],
-                    "runtime": {
-                        "cross_validation": 10,
-                        "stratified": False
-                    },
-                    "inputs": ["cast_1_step", "extract_target_step"]
-                }
-            ]
-        }
-
-    # @override
-
-    def importance(datset, problem_description):
-        return 7
-
-
-class UU3TestTemplate(DSBoxTemplate):
-    def __init__(self):
-        DSBoxTemplate.__init__(self)
-        self.template = {
-            "name": "UU3_Test_Template",
-            "taskSubtype": {TaskSubtype.UNIVARIATE.name, TaskSubtype.MULTIVARIATE.name},
-            "taskType": TaskType.REGRESSION.name,
-            "inputType": "table",
-            "output": "model_step",  # Name of the final step generating the prediction
-            "target": "extract_target_step",  # Name of the step generating the ground truth
-            "steps": [
-                {
-                    "name": "multi_table_processing_step",
-                    "primitives": ["d3m.primitives.dsbox.MultiTableFeaturization"],
-                    "inputs": ["template_input"]
-                },
-                {
-                    "name": "extract_attribute_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
-                        "hyperparameters":
-                            {
-                                'semantic_types': ('https://metadata.datadrivendiscovery.org/types/Attribute',),
-                                'use_columns': (),
-                                'exclude_columns': ()
-                            }
-                    }],
-                    "inputs": ["multi_table_processing_step"]
-                },
-                {
-                    "name": "extract_target_step",
-                    "primitives": [{
-                        "primitive": "d3m.primitives.data.ExtractColumnsBySemanticTypes",
-                        "hyperparameters":
-                            {
-                                'semantic_types': ('https://metadata.datadrivendiscovery.org/types/TrueTarget',),
-                                'use_columns': (),
-                                'exclude_columns': ()
-                            }
-                        }],
-                    "inputs": ["multi_table_processing_step"]
-                },
-                {
-                    "name": "encode1_step",
-                    "primitives": ["d3m.primitives.dsbox.UnaryEncoder"],
-                    "inputs": ["extract_attribute_step"]
-                },
-                {
-                    "name": "encode2_step",
-                    "primitives": ["d3m.primitives.dsbox.Encoder"],
-                    "inputs": ["encode1_step"]
-                },
-                {
-                    "name": "impute_step",
-                    "primitives": ["d3m.primitives.dsbox.MeanImputation"],
-                    "inputs": ["encode2_step"]
-                },
-                {
-                    "name": "cast_step",
-                    "primitives": [
-                        {
-                            "primitive": "d3m.primitives.data.CastToType",
-                            "hyperparameters": {"type_to_cast": ["float"]}
-                        },
-                        {
-                            "primitive": "d3m.primitives.dsbox.DoNothing",
-                            "hyperparameters": {}
-                        }
-                    ],
-                    "inputs": ["impute_step"]
-                },
-                # {
-                #     "name": "corex_step",
-                #     "primitives": ["d3m.primitives.dsbox.CorexText"],
-                #     "inputs": ["cast_1_step"]
-                # },
-                {
-                    "name": "model_step",
-                    "primitives": [{
-                        "primitive":
-                            "d3m.primitives.sklearn_wrap.SKRandomForestRegressor",
-                        "hyperparameters":
-                            {
-                            }
-                    }
-                    ],
-                    "inputs": ["cast_step", "extract_target_step"]
-                }
-            ]
-        }
-
-    # @override
-    def importance(datset, problem_description):
-        return 7
